AMMI.contour {agricolae}	R Documentation

AMMI contour

Description

Draws a polygon or a circumference around the center of the Biplot with a proportional radio at the longest distance of the genotype.

Usage

AMMI.contour(model, distance, shape, ...)

Arguments

model
Object

distance
Circumference radius >0 and <=1

shape
Numerical, relating to the shape of the polygon outline.

...
Parameters corresponding to the R lines function

Details

First, it is necessary to execute the AMMI function. It is only valid for the BIPLOT function but not for the TRIPLOT one.

Value

Genotypes within and outside the area.

distance
Distance from genotype to origin (0,0)

Note

Complement graphics AMMI

Author(s)

Felipe de Mendiburu

See Also

AMMI

Examples

library(agricolae)
# see AMMI.
data(sinRepAmmi)
Environment <- sinRepAmmi$ENV
Genotype <- sinRepAmmi$GEN
Yield <- sinRepAmmi$YLD
REP <- 3
MSerror <- 93.24224
model<-AMMI(Environment, Genotype, REP, Yield, MSerror)
plot(model)
AMMI.contour(model,distance=0.7,shape=8,col="red",lwd=2,lty=5)

[Package agricolae version 1.3-7 Index]