CIC {agricolae}	R Documentation

Data for late blight of potatoes

Description

A study of Phytophthora infestans in the potato plant in the localities of Comas and Oxapampa in Peru, 2005.

Usage

data(CIC)

Format

The format is: List of 2 (comas, oxapampa)

Locality

a factor with levels Comas Oxapampa

Genotype

a factor

Rep

a numeric vector, replications

E9

a numeric vector, infestans percentaje to 9 days

AUDPC

a numeric vector: the area under the disease-progress curve

Relative

a numeric vector, relative area

Details

comas: temperature=59.9 Fahrenheit, relative humidity=83.3 oxapampa: temperature=64.8 Fahrenheit, relative humidity=86.2 AUDPC and relative see function audpc(). help(audpc) Exx: Evaluation in percentaje, xx is days. ORD1, ORD2, SBLK and row are references location of the plot in the field.

Source

Experimental field, 2004-2005. Data Kindly provided by Matilde Orrillo.

References

International Potato Center. CIP - Lima Peru.

Examples

library(agricolae)
data(CIC)
CIC$comas
CIC$oxapampa

[Package agricolae version 1.3-7 Index]