Chz2006 {agricolae}	R Documentation

Data amendment Carhuaz 2006

Description

Incidents and performance of healthy tubers and rotten potato field infested with naturally Ralstonia solanacearum Race 3/Bv 2A, after application of inorganic amendments and a rotation crop in Carhuaz Peru, 2006.

Usage

data(Chz2006)

Format

The format is: List of 2

amendment

a factor

crop

a factor

block

a numeric vector, replications

plant

a numeric vector, number plant

wilt_percent

a numeric vector, wilt percentage at 60 days

health

a numeric vector, kg/8m2

rot

a numeric vector, kg/8m2

Details

Application of inorganic amendment and crop rotation to control bacterial wilt of the potato (MBP).

Source

Experimental field, 2006. Data Kindly provided by Pedro Aley.

References

International Potato Center. CIP - Lima Peru.

Examples

library(agricolae)
data(Chz2006)
str(Chz2006)
wilt<-Chz2006$wilt
yield<-Chz2006$yield
means <- tapply.stat(wilt[,5],wilt[,1:3],function(x) mean(x,na.rm=TRUE))
names(means)[4]<-"wilt_percent"
model <- aov(wilt_percent ~ block + crop, means)
anova(model)
cv.model(model)
yield<-yield[order(paste(yield[,1],yield[,2],yield[,3])),]
correlation(means[,4],yield[,4],method="spearman")

[Package agricolae version 1.3-7 Index]