DAU.test {agricolae}	R Documentation

Finding the Variance Analysis of the Augmented block Design

Description

Analysis of variance Augmented block and comparison mean adjusted.

Usage

DAU.test(block, trt, y, method = c("lsd","tukey"),alpha=0.05,group=TRUE,console=FALSE)

Arguments

block
blocks

trt
Treatment

y
Response

method
Comparison treatments

alpha
Significant test

group
TRUE or FALSE

console
logical, print output

Details

Method of comparison treatment. lsd: Least significant difference. tukey: Honestly significant differente. The controls can have different repetitions, at least two, do not use missing data.

Value

means
Statistical summary of the study variable

parameters
Design parameters

statistics
Statistics of the model

comparison
Comparison between treatments

groups
Formation of treatment groups

SE.difference
Standard error of:
Two Control Treatments
Two Augmented Treatments
Two Augmented Treatments(Different Blocks)
A Augmented Treatment and A Control Treatment

vartau
Variance-covariance matrix of the difference in treatments

Author(s)

F. de Mendiburu

References

Federer, W. T. (1956). Augmented (or hoonuiaku) designs. Hawaiian Planters, Record LV(2):191-208.

See Also

BIB.test, duncan.test, durbin.test, friedman, HSD.test, kruskal, LSD.test, Median.test, PBIB.test, REGW.test, scheffe.test, SNK.test, waerden.test, waller.test, plot.group

Examples

library(agricolae)
block<-c(rep("I",7),rep("II",6),rep("III",7))
trt<-c("A","B","C","D","g","k","l","A","B","C","D","e","i","A","B","C","D","f","h","j")
yield<-c(83,77,78,78,70,75,74,79,81,81,91,79,78,92,79,87,81,89,96,82)
out<- DAU.test(block,trt,yield,method="lsd", group=TRUE)
print(out$groups)
plot(out)

[Package agricolae version 1.3-7 Index]