Data amendment Huanuco 2006

Incidents and performance of healthy tubers and rotten potato field infested with naturally Ralstonia solanacearum Race 3/Bv 2A, after application of inorganic amendments and a rotation crop in Huanuco Peru, 2006.

Format

The format is: List of 2

list("amendment")

a factor

list("crop")

a factor

list("block")

a numeric vector, replications

list("plant")

a numeric vector, number plant

list("wilt_percent")

a numeric vector, wilt percentage at 60 days

list("health")

a numeric vector, kg/8m2, 20 plants

list("rot")

a numeric vector, kg/8m2, 20 plants

Source

Experimental field, 2006. Data Kindly provided by Pedro Aley.

Details

Application of inorganic amendment and crop rotation to control bacterial wilt of the potato (MBP).

References

International Potato Center. CIP - Lima Peru.

Examples

library(agricolae)
data(Hco2006)
str(Hco2006)

#> List of 2
#>  $ wilt :'data.frame':	1347 obs. of  5 variables:
#>   ..$ amendment   : Factor w/ 3 levels "0C","2C1Z","4C": 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ crop        : Factor w/ 3 levels "Cabbage","Corn",..: 3 3 3 3 3 3 3 3 3 3 ...
#>   ..$ block       : Factor w/ 3 levels "I","II","III": 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ plant       : num [1:1347] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ wilt_percent: num [1:1347] 0 0 0 0 50 0 100 0 0 0 ...
#>  $ yield:'data.frame':	27 obs. of  5 variables:
#>   ..$ amendment: Factor w/ 3 levels "0C","2C1Z","4C": 1 1 1 1 1 1 1 1 1 3 ...
#>   ..$ crop     : Factor w/ 3 levels "Cabbage","Corn",..: 3 3 3 1 1 1 2 2 2 3 ...
#>   ..$ block    : Factor w/ 3 levels "I","II","III": 1 2 3 1 2 3 1 2 3 1 ...
#>   ..$ health   : num [1:27] 3.9 11.4 13.1 8.2 13.1 14.5 4.5 13.4 14.4 3.8 ...
#>   ..$ rot      : num [1:27] 1.5 2.6 2.8 1.5 2.15 0.5 1.3 3.6 4.7 0.8 ...

wilt<-Hco2006$wilt
yield<-Hco2006$yield
means <- tapply.stat(wilt[,5],wilt[,1:3],function(x) mean(x,na.rm=TRUE))
names(means)[4]<-"wilt_percent"
model <- aov(wilt_percent ~ block + crop, means)
anova(model)

#> Analysis of Variance Table
#> 
#> Response: wilt_percent
#>           Df Sum Sq Mean Sq F value Pr(>F)
#> block      2  569.9  284.93  1.2056 0.3186
#> crop       2  496.1  248.04  1.0495 0.3670
#> Residuals 22 5199.5  236.34               

cv.model(model)

#> [1] 69.41147

yield<-yield[order(paste(yield[,1],yield[,2],yield[,3])),]
correlation(means[,4],yield[,4],method="spearman")

#> 
#> Spearman's rank correlation rho
#> 
#> data: means[, 4] and yield[, 4] 
#> p-value = 0.006544646 
#> alternative hypothesis: true rho is not equal to 0
#> sample estimates:
#> rho
#>  -0.5102417