REGW.test {agricolae}	R Documentation

Ryan, Einot and Gabriel and Welsch multiple range test

Description

Multiple range tests for all pairwise comparisons, to obtain a confident inequalities multiple range tests.

Usage

REGW.test(y, trt, DFerror, MSerror, alpha = 0.05, group=TRUE, main = NULL,console=FALSE)

Arguments

y
model(aov or lm) or answer of the experimental unit

trt
Constant( only y=model) or vector treatment applied to each experimental unit

DFerror
Degree free

MSerror
Mean Square Error

alpha
Significant level

group
TRUE or FALSE

main
Title

console
logical, print output

Details

It is necessary first makes a analysis of variance.

if y = model, then to apply the instruction:
REGW.test (model, "trt", alpha = 0.05, group = TRUE, main = NULL, console = FALSE)
where the model class is aov or lm.

Value

statistics
Statistics of the model

parameters
Design parameters

regw
Critical Range Table

means
Statistical summary of the study variable

comparison
Comparison between treatments

groups
Formation of treatment groups

Author(s)

Felipe de Mendiburu

References

Multiple comparisons theory and methods. Departament of statistics the Ohio State University. USA, 1996. Jason C. Hsu. Chapman Hall/CRC

See Also

BIB.test, DAU.test, duncan.test, durbin.test, friedman, HSD.test, kruskal, LSD.test, Median.test, PBIB.test, scheffe.test, SNK.test, waerden.test, waller.test, plot.group

Examples

library(agricolae)
data(sweetpotato)
model<-aov(yield~virus,data=sweetpotato)
out<- REGW.test(model,"virus", 
main="Yield of sweetpotato. Dealt with different virus")
print(out)
REGW.test(model,"virus",alpha=0.05,console=TRUE,group=FALSE)

[Package agricolae version 1.3-7 Index]