SNK.test {agricolae}	R Documentation

Student-Newman-Keuls (SNK)

Description

SNK is derived from Tukey, but it is less conservative (finds more differences). Tukey controls the error for all comparisons, where SNK only controls for comparisons under consideration. The level by alpha default is 0.05.

Usage

SNK.test(y, trt, DFerror, MSerror, alpha = 0.05, group=TRUE, main = NULL,console=FALSE)

Arguments

y
model(aov or lm) or answer of the experimental unit

trt
Constant( only y=model) or vector treatment applied to each experimental unit

DFerror
Degree free

MSerror
Mean Square Error

alpha
Significant level

group
TRUE or FALSE

main
Title

console
logical, print output

Details

It is necessary first makes a analysis of variance.

if y = model, then to apply the instruction:
SNK.test (model, "trt", alpha = 0.05, group = TRUE, main = NULL, console = FALSE)
where the model class is aov or lm.

Value

statistics
Statistics of the model

parameters
Design parameters

snk
Critical Range Table

means
Statistical summary of the study variable

comparison
Comparison between treatments

groups
Formation of treatment groups

Author(s)

Felipe de Mendiburu

References

1. Principles and procedures of statistics a biometrical approach Steel & Torry & Dickey. Third Edition 1997
2. Multiple comparisons theory and methods. Departament of statistics the Ohio State University. USA, 1996. Jason C. Hsu. Chapman Hall/CRC.

See Also

BIB.test, DAU.test, duncan.test, durbin.test, friedman, HSD.test, kruskal, LSD.test, Median.test, PBIB.test, REGW.test, scheffe.test, waerden.test, waller.test, plot.group

Examples

library(agricolae)
data(sweetpotato)
model<-aov(yield~virus,data=sweetpotato)
out <- SNK.test(model,"virus", console=TRUE, 
main="Yield of sweetpotato. Dealt with different virus")
print(SNK.test(model,"virus", group=FALSE))
# version old SNK.test()
df<-df.residual(model)
MSerror<-deviance(model)/df
out <- with(sweetpotato,SNK.test(yield,virus,df,MSerror, group=TRUE))
print(out$groups)

[Package agricolae version 1.3-7 Index]