waerden.test {agricolae}	R Documentation

Multiple comparisons. The van der Waerden (Normal Scores)

Description

A nonparametric test for several independent samples.

Usage

waerden.test(y, trt, alpha=0.05, group=TRUE, main=NULL,console=FALSE)

Arguments

y
Variable response

trt
Treatments

alpha
Significant level

group
TRUE or FALSE

main
Title

console
logical, print output

Details

The data consist of k samples of possibly unequal sample size.
The post hoc test is using the criterium Fisher's least
significant difference (LSD).

Value

statistics
Statistics of the model

parameters
Design parameters

means
Statistical summary of the study variable

comparison
Comparison between treatments

groups
Formation of treatment groups

Author(s)

Felipe de Mendiburu

References

Practical Nonparametrics Statistics. W.J. Conover, 1999

See Also

BIB.test, DAU.test, duncan.test, durbin.test, friedman, HSD.test, kruskal, LSD.test, Median.test, PBIB.test, REGW.test, scheffe.test, SNK.test, waller.test, plot.group

Examples

library(agricolae)
# example 1
data(corn)
out1<-with(corn,waerden.test(observation,method,group=TRUE))
print(out1$groups)
plot(out1)
out2<-with(corn,waerden.test(observation,method,group=FALSE))
print(out2$comparison)
# example 2
data(sweetpotato)
out<-with(sweetpotato,waerden.test(yield,virus,alpha=0.01,group=TRUE))
print(out)

[Package agricolae version 1.3-7 Index]