bar.group {agricolae}	R Documentation

Plotting the multiple comparison of means

Description

It plots bars of the averages of treatments to compare. It uses the objects generated by a procedure of comparison like LSD, HSD, Kruskall, Waller-Duncan, Friedman or Durbin. It can also display the 'average' value over each bar in a bar chart.

Usage

bar.group(x,horiz=FALSE, decreasing=TRUE, ...)

Arguments

x
Object created by a test of comparison

horiz
Horizontal or vertical bars

decreasing
Logical, decreasing order of the mean

...
Parameters of the function barplot()

Details

x: data frame formed by 5 columns: name of the bars, height and level of the bar.

Value

A list with numeric vectors giving the coordinates of all the bar midpoints drawn.

x
eje-1 coordinate

height
eje-2 coordinate by group

Author(s)

Felipe de Meniburu

See Also

LSD.test, HSD.test, kruskal , friedman, durbin.test, waller.test , plot.group

Examples

# Example 1
library(agricolae)
data(sweetpotato)
model<-aov(yield~virus,data=sweetpotato)
comparison<- LSD.test(model,"virus",alpha=0.01,group=TRUE)
print(comparison$groups)
oldpar<-par(cex=1.5)
bar.group(comparison$groups,horiz=TRUE,density=8,col="blue",border="red", xlim=c(0,50),las=1)
title(cex.main=0.8,main="Comparison between\ntreatment means",xlab="Yield",ylab="Virus")
# Example 2
library(agricolae)
x <- 1:4
y <- c(0.29, 0.44, 0.09, 0.49)
xy <- data.frame(x,y,y)
par(oldpar)
oldpar<-par(cex=1.5)
bar.group(xy,density=30,angle=90,col="brown",border=FALSE,ylim=c(0,0.6),lwd=2,las=1)
par(oldpar)

[Package agricolae version 1.3-7 Index]