cv.model {agricolae}	R Documentation

Coefficient of the experiment variation

Description

It obtains the coefficient of variation of the experiment obtained by models lm() or aov()

Usage

cv.model(x)

Arguments

x
object of model lm() or AOV()

Details

sqrt(MSerror)*100/mean(x)

Value

Returns the coefficient of variation of the experiment according to the applied statistical model

Author(s)

Felipe de Mendiburu

See Also

LSD.test, HSD.test, waller.test

Examples

# see examples from LSD , Waller-Duncan or HSD and complete with it:
library(agricolae)
# not run
# cv<-cv.model(model)

[Package agricolae version 1.3-7 Index]