cv.similarity {agricolae}	R Documentation

Coefficient of the similarity matrix variation

Description

This process consists of finding the coefficient of the distances of similarity of binary tables (1 and 0) as used for scoring molecular marker data for presence and absence of PCR amplification products.

Usage

cv.similarity(A)

Arguments

A
matrix of binary data

Value

Returns the coefficient of variation of the similarity model

Author(s)

Felipe de Mendiburu

See Also

similarity, resampling.cv

Examples

# molecular markers.
library(agricolae)
data(markers)
cv<-cv.similarity(markers)

[Package agricolae version 1.3-7 Index]