design.ab {agricolae}	R Documentation

Design of experiments for a factorial

Description

It generates a design of blocks, randomize and latin square for combined n. factors uses the methods of number generation in R. The seed is by set.seed(seed, kinds).

Usage

design.ab(trt, r, serie = 2, design=c("rcbd","crd","lsd"),
seed = 0, kinds = "Super-Duper",first=TRUE,randomization=TRUE)

Arguments

trt
n levels factors

r
Replications or Blocks

serie
number plot, 1: 11,12; 2: 101,102; 3: 1001,1002

design
type

seed
Seed

kinds
Method for to randomize

first
TRUE or FALSE - randomize rep 1

randomization
TRUE or FALSE - randomize

Details

kinds <- c("Wichmann-Hill", "Marsaglia-Multicarry", "Super-Duper", "Mersenne-Twister", "Knuth-TAOCP", "user-supplied", "Knuth-TAOCP-2002", "default" )

Value

parameters
Design parameters

book
Fieldbook

Author(s)

Felipe de Mendiburu

References

Introduction to Experimental Statistics. Ching Chun Li. McGraw-Hill Book Company, INC, New. York, 1964

See Also

design.split, design.alpha,design.bib, design.crd , design.cyclic , design.dau , design.graeco, design.lattice, design.lsd, design.rcbd, design.strip

Examples

# factorial 3 x 2 with 3 blocks
library(agricolae)
trt<-c(3,2) # factorial 3x2
outdesign <-design.ab(trt, r=3, serie=2)
book<-outdesign$book
head(book,10) # print of the field book
# factorial 2 x 2 x 2 with 5 replications in completely randomized design.
trt<-c(2,2,2)
outdesign<-design.ab(trt, r=5, serie=2,design="crd")
book<-outdesign$book
print(book)
# factorial 3 x 3 in latin square design.
trt <-c(3,3)
outdesign<-design.ab(trt, serie=2, design="lsd")
book<-outdesign$book
print(book)

[Package agricolae version 1.3-7 Index]