design.alpha {agricolae}	R Documentation

Alpha design type (0,1)

Description

Generates an alpha designs starting from the alpha design fixing under the series formulated by Patterson and Williams. These designs are generated by the alpha arrangements. They are similar to the lattice designs, but the tables are rectangular s by k (with s blocks and k<s columns. The number of treatments should be equal to s*k and all the experimental units r*s*k (r replications).

Usage

design.alpha(trt, k, r, serie = 2, seed = 0, kinds = "Super-Duper",randomization=TRUE)

Arguments

trt
Treatments

k
size block

r
Replications

serie
number plot, 1: 11,12; 2: 101,102; 3: 1001,1002

seed
seed

kinds
method for to randomize

randomization
TRUE or FALSE - randomize

Details

Parameters for the alpha design: I. r=2, k <= s; II. r=3, s odd, k <= s; III.r=3, s even, k <= s-1; IV. r=4, s odd but not a multiple of 3, k<=s

r= replications s=number of blocks k=size of block Number of treatment is equal to k*s

Value

parameters
Design parameters

statistics
Design statistics

sketch
Design sketch

book
Fieldbook

Author(s)

Felipe de Mendiburu

References

H.D. Patterson and E.R. Williams. Biometrika (1976) A new class of resolvable incomplete block designs. printed in Great Britain. Online: http://biomet.oxfordjournals.org/cgi/content/abstract/63/1/83

See Also

design.ab, design.split,design.bib, design.crd , design.cyclic , design.dau , design.graeco, design.lattice, design.lsd, design.rcbd, design.strip

Examples

library(agricolae)
#Example one
trt<-1:30
t <- length(trt)
# size block k
k<-3
# Blocks s
s<-t/k
# replications r
r <- 2
outdesign<- design.alpha(trt,k,r,serie=2)
book<-outdesign$book
plots<-book[,1]
dim(plots)<-c(k,s,r)
for (i in 1:r) print(t(plots[,,i]))
outdesign$sketch
# Example two 
trt<-letters[1:12] 
t <- length(trt)
k<-3
r<-3
s<-t/k
outdesign<- design.alpha(trt,k,r,serie=2)
book<-outdesign$book
plots<-book[,1]
dim(plots)<-c(k,s,r)
for (i in 1:r) print(t(plots[,,i]))
outdesign$sketch

[Package agricolae version 1.3-7 Index]