design.cyclic {agricolae}	R Documentation

Cyclic designs

Description

The cyclic design is a incomplete blocks designs, it is generated from a incomplete block initial of the size k, the plan is generated and randomized. The efficient and robust cyclic designs for 6 to 30 treatments, replications <= 10.

Usage

design.cyclic(trt, k, r, serie = 2, rowcol = FALSE, seed = 0, kinds = "Super-Duper"
,randomization=TRUE)

Arguments

trt
vector treatments

k
block size

r
Replications

serie
number plot, 1: 11,12; 2: 101,102; 3: 1001,1002

rowcol
TRUE: row-column design

seed
init seed random

kinds
random method

randomization
TRUE or FALSE - randomize

Details

Number o treatment 6 to 30. (r) Replication 2 to 10. (k) size of block 2 to 10. replication = i*k, "i" is value integer.

Value

parameters
Design parameters

sketch
Design sketch

book
Fieldbook

Author(s)

Felipe de Mendiburu

References

Kuehl, Robert(2000), Design of Experiments. 2nd ed., Duxbury. John, J.A. (1981) Efficient Cyclic Design. J. R. Statist. Soc. B, 43, No. 1, pp, 76-80.

See Also

design.ab, design.alpha,design.bib, design.crd , design.split , design.dau , design.graeco, design.lattice, design.lsd, design.rcbd, design.strip

Examples

library(agricolae)
trt<-letters[1:8]
# block size = 2, replication = 6
outdesign1 <- design.cyclic(trt,k=2, r=6,serie=2)
names(outdesign1)
# groups 1,2,3
outdesign1$sketch[[1]]
outdesign1$sketch[[2]]
outdesign1$sketch[[3]]
outdesign1$book
# row-column design
outdesign2<- design.cyclic(trt,k=2, r=6, serie=2, rowcol=TRUE)
outdesign2$sketch

[Package agricolae version 1.3-7 Index]