Graeco - latin square design

A graeco - latin square is a KxK pattern that permits the study of k treatments simultaneously with three different blocking variables, each at k levels.

The function is only for squares of the odd numbers and even numbers (4, 8, 10 and 12)

design.graeco(trt1, trt2, serie = 2, seed = 0, kinds = "Super-Duper",randomization=TRUE)

Arguments

trt1
Treatments

trt2
Treatments

serie
number plot, 1: 11,12; 2: 101,102; 3: 1001,1002

seed
seed

kinds
method for to randomize

randomization
TRUE or FALSE - randomize

Details

kinds <- c("Wichmann-Hill", "Marsaglia-Multicarry", "Super-Duper", "Mersenne-Twister", "Knuth-TAOCP", "user-supplied", "Knuth-TAOCP-2002", "default" )

Value

parameters

Design parameters

book

Fieldbook

References

1. Statistics for Experimenters Design, Innovation, and Discovery Second Edition. George E. P. Box. Wiley-Interscience. 2005.

2. Experimental design. Cochran and Cox. Second edition. Wiley Classics Library Edition published 1992.

See also

Examples

library(agricolae)
T1<-c("a","b","c","d")
T2<-c("v","w","x","y")
outdesign <- design.graeco(T1,T2,serie=1)
graeco<-outdesign$book
plots <-as.numeric(graeco[,1])
print(outdesign$sketch)

#>      [,1]  [,2]  [,3]  [,4] 
#> [1,] "d x" "b v" "c w" "a y"
#> [2,] "b w" "d y" "a x" "c v"
#> [3,] "c y" "a w" "d v" "b x"
#> [4,] "a v" "c x" "b y" "d w"

print(matrix(plots,byrow=TRUE,ncol=4))

#>      [,1] [,2] [,3] [,4]
#> [1,]   11   12   13   14
#> [2,]   21   22   23   24
#> [3,]   31   32   33   34
#> [4,]   41   42   43   44

# 10 x 10
T1 <- letters[1:10]
T2 <- 1:10
outdesign <-  design.graeco(T1,T2,serie=2)
print(outdesign$sketch)

#>       [,1]   [,2]   [,3]   [,4]   [,5]   [,6]   [,7]   [,8]   [,9]   [,10] 
#>  [1,] "c 10" "f 4"  "i 5"  "b 1"  "e 8"  "a 2"  "h 6"  "j 7"  "d 3"  "g 9" 
#>  [2,] "h 1"  "i 3"  "g 4"  "e 5"  "b 10" "j 8"  "a 7"  "f 6"  "c 9"  "d 2" 
#>  [3,] "a 6"  "h 10" "e 9"  "d 4"  "j 5"  "b 3"  "f 8"  "g 1"  "i 2"  "c 7" 
#>  [4,] "g 8"  "a 1"  "h 3"  "j 2"  "c 4"  "f 5"  "b 9"  "d 10" "e 7"  "i 6" 
#>  [5,] "b 2"  "d 8"  "a 10" "h 9"  "f 7"  "i 4"  "g 5"  "c 3"  "j 6"  "e 1" 
#>  [6,] "d 5"  "b 7"  "c 8"  "a 3"  "h 2"  "g 6"  "e 4"  "i 9"  "f 1"  "j 10"
#>  [7,] "j 4"  "c 5"  "b 6"  "i 8"  "a 9"  "h 7"  "d 1"  "e 2"  "g 10" "f 3" 
#>  [8,] "i 7"  "e 6"  "j 1"  "f 10" "g 3"  "d 9"  "c 2"  "b 4"  "h 5"  "a 8" 
#>  [9,] "e 3"  "j 9"  "f 2"  "g 7"  "d 6"  "c 1"  "i 10" "h 8"  "a 4"  "b 5" 
#> [10,] "f 9"  "g 2"  "d 7"  "c 6"  "i 1"  "e 10" "j 3"  "a 5"  "b 8"  "h 4"