design.youden {agricolae}	R Documentation

Incomplete Latin Square Design

Description

Such designs are referred to as Youden squares since they were introduced by Youden (1937) after Yates (1936) considered the special case of column equal to number treatment minus 1. "Random" uses the methods of number generation in R. The seed is by set.seed(seed, kinds).

Usage

design.youden(trt, r, serie = 2, seed = 0, kinds = "Super-Duper",first=TRUE
,randomization=TRUE)

Arguments

trt
Treatments

r
Replications or number of columns

serie
number plot, 1: 11,12; 2: 101,102; 3: 1001,1002

seed
seed

kinds
method for to randomize

first
TRUE or FALSE - randomize rep 1

randomization
TRUE or FALSE - randomize

Details

kinds <- c("Wichmann-Hill", "Marsaglia-Multicarry", "Super-Duper", "Mersenne-Twister", "Knuth-TAOCP", "user-supplied", "Knuth-TAOCP-2002", "default" )

Value

parameters
Design parameters

sketch
Design sketch

book
Fieldbook

Author(s)

Felipe de Mendiburu

References

Design and Analysis of experiment. Hinkelmann, Klaus and Kempthorne, Oscar. Wiley-Interscience. Copyright (2008) by John Wiley and Sons. Inc., Hoboken, new Yersy

See Also

design.ab, design.alpha,design.bib, design.crd , design.cyclic , design.dau , design.graeco, design.lattice, design.split, design.rcbd, design.strip, design.lsd

Examples

library(agricolae)
varieties<-c("perricholi","yungay","maria bonita","tomasa")
r<-3
outdesign <-design.youden(varieties,r,serie=2,seed=23)
youden <- outdesign$book
print(outdesign$sketch)
plots <-as.numeric(youden[,1])
print(matrix(plots,byrow=TRUE,ncol=r))
print(youden) # field book.
# Write on hard disk.
# write.table(youden,"youden.txt", row.names=FALSE, sep="\t")
# file.show("youden.txt")

[Package agricolae version 1.3-7 Index]