design.mat {agricolae}	R Documentation

Experimental design matrix

Description

Generate the design matrix from the fieldbook generated by an experimental plan or a dataframe for analysis.

Usage

design.mat(book, locations)

Arguments

book
data frame or matrix, field book

locations
numeric, column position of the field book

Value

X is matrix design.

Author(s)

Felipe de Mendiburu

See Also

design.ab, design.alpha,design.bib, design.crd , design.cyclic , design.split , design.graeco, design.lattice, design.lsd, design.rcbd, design.strip, design.dau

Examples

# dataframe: data analysis
library(agricolae)
data(sweetpotato)
X<-design.mat(sweetpotato,1)
print(X)
# fieldbook: RCBD design
trt <- LETTERS[1:4]
r<-3
plan<-design.rcbd(trt,r,seed=11)
X<-design.mat(plan$book,2:3)
print(X)

[Package agricolae version 1.3-7 Index]