friedman {agricolae}	R Documentation

Friedman test and multiple comparison of treatments

Description

The data consist of b-blocks mutually independent k-variate random variables Xij, i=1,..,b; j=1,..k. The random variable X is in block i and is associated with treatment j. It makes the multiple comparison of the Friedman test with or without ties. A first result is obtained by friedman.test of R.

Usage

friedman(judge,trt,evaluation,alpha=0.05,group=TRUE,main=NULL,console=FALSE)

Arguments

judge
Identification of the judge in the evaluation

trt
Treatment

evaluation
Variable

alpha
Significant test

group
TRUE or FALSE

main
Title

console
logical, print output

Details

The post hoc friedman test is using the criterium Fisher's least significant difference (LSD)

Value

statistics
Statistics of the model

parameters
Design parameters

means
Statistical summary of the study variable

comparison
Comparison between treatments

groups
Formation of treatment groups

Author(s)

Felipe de Mendiburu

References

Practical Nonparametrics Statistics. W.J. Conover, 1999

See Also

BIB.test, DAU.test, duncan.test, durbin.test, HSD.test, kruskal, LSD.test, Median.test, PBIB.test, REGW.test, scheffe.test, SNK.test, waerden.test, waller.test, plot.group

Examples

library(agricolae)
data(grass)
out<-with(grass,friedman(judge,trt, evaluation,alpha=0.05, group=TRUE,console=TRUE,
main="Data of the book of Conover"))
#startgraph
plot(out,variation="IQR")
#endgraph

[Package agricolae version 1.3-7 Index]