graph.freq {agricolae}	R Documentation

Histogram

Description

In many situations it has intervals of class defined with its respective frequencies. By means of this function, the graphic of frequency is obtained and it is possible to superpose the normal distribution, polygon of frequency, Ojiva and to construct the table of complete frequency.

Usage

graph.freq(x, breaks=NULL,counts=NULL,frequency=1, plot=TRUE, nclass=NULL,
xlab="",ylab="",axes = "",las=1,...)

Arguments

x
a vector of values, a object hist(), graph.freq()

counts
frequency and x is class intervals

breaks
a vector giving the breakpoints between histogram cells

frequency
1=counts, 2=relative, 3=density

plot
logic

nclass
number of classes

xlab
x labels

ylab
y labels

las
values 0,1,2 and 3 are the axis styles. see plot()

axes
TRUE or FALSE

...
other parameters of plot

Value

breaks
a vector giving the breakpoints between histogram cells

counts
frequency and x is class intervals

mids
center point in class

relative
Relative frequency, height

density
Density frequency, height

Author(s)

Felipe de Mendiburu

See Also

polygon.freq, table.freq, stat.freq,inter.freq,sturges.freq, join.freq,ogive.freq, normal.freq

Examples

library(agricolae)
data(genxenv)
yield <- subset(genxenv$YLD,genxenv$ENV==2)
yield <- round(yield,1)
h<- graph.freq(yield,axes=FALSE, frequency=1, ylab="frequency",col="yellow")
axis(1,h$breaks)
axis(2,seq(0,20,0.1))
# To reproduce histogram.
h1 <- graph.freq(h, col="blue", frequency=2,border="red", density=8,axes=FALSE,
xlab="YIELD",ylab="relative")
axis(1,h$breaks)
axis(2,seq(0,.4,0.1))
# summary, only frecuency
limits <-seq(10,40,5)
frequencies <-c(2,6,8,7,3,4)
#startgraph
h<-graph.freq(limits,counts=frequencies,col="bisque",xlab="Classes")
polygon.freq(h,col="red")
title( main="Histogram and polygon of frequency",
ylab="frequency")
#endgraph
# Statistics
measures<-stat.freq(h)
print(measures)
# frequency table full
round(table.freq(h),2)
#startgraph
# ogive
ogive.freq(h,col="red",type="b",ylab="Accumulated relative frequency",
xlab="Variable")
# only .frequency polygon
h<-graph.freq(limits,counts=frequencies,border=FALSE,col=NULL,xlab="  ",ylab="")
title( main="Polygon of frequency",
xlab="Variable", ylab="Frecuency")
polygon.freq(h,col="blue")
grid(col="brown")
#endgraph
# Draw curve for Histogram
h<- graph.freq(yield,axes=FALSE, frequency=3, ylab="f(yield)",col="yellow")
axis(1,h$breaks)
axis(2,seq(0,0.18,0.03),las=2)
lines(density(yield), col = "red", lwd = 2)
title("Draw curve for Histogram")

[Package agricolae version 1.3-7 Index]