Data of AUDPC for nonparametrical stability analysis

Published data. Haynes. Mean area under the disease progress curve (AUDPC) for each of 16 potato clones evaluated at eight sites across the United States in 1996

data(haynes)

Format

A data.frame with 16 observations on the following 9 variables.

@details

clone a factor with levels A84118-3, AO80432-1, AO84275-3, AWN86514-2, B0692-4, B0718-3, B0749-2F, B0767-2, Bertita, Bzura, C0083008-1, Elba, Greta, Krantz, Libertas, Stobrawa

FL FL

ME ME

MI MI

MN MN

ND ND

NY NY

PA PA

WI WI

References

Haynes K G, Lambert D H, Christ B J, Weingartner D P, Douches D S, Backlund J E, Fry W and Stevenson W. 1998. Phenotypic stability of resistance to late blight in potato clones evaluated at eight sites in the United States American Journal Potato Research 75, pag 211-217.

Examples

library(agricolae)
data(haynes)
str(haynes)

#> 'data.frame':	16 obs. of  9 variables:
#>  $ clone: Factor w/ 16 levels "A84118-3","AO80432-1",..: 1 2 3 4 5 6 7 8 9 10 ...
#>  $ FL   : int  284 254 395 136 87 130 571 141 309 320 ...
#>  $ MI   : int  1113 690 1089 296 653 126 1212 191 649 300 ...
#>  $ ME   : int  1053 1112 1090 374 412 329 1323 258 929 713 ...
#>  $ MN   : int  997 832 387 44 73 72 601 93 325 155 ...
#>  $ ND   : int  612 761 633 377 334 444 908 327 648 554 ...
#>  $ NY   : int  590 518 296 101 70 167 579 64 302 242 ...
#>  $ PA   : int  484 555 347 46 26 61 498 79 290 123 ...
#>  $ WI   : int  800 1153 850 41 141 211 491 71 565 247 ...