hcut {agricolae}	R Documentation

Cut tree of consensus

Description

It shows dendrogram of a consensus of a tree generated by hclust.

Usage

hcut(consensus,h,group,col.text="blue",cex.text=1,...)

Arguments

consensus
object consensus

h
numeric scalar or vector with heights where the tree should be cut.

group
an integer scalar with the desired number of group

col.text
color of number consensus

cex.text
size of number consensus

...
Other parameters of the function plot() in cut()

Value

hcut Returns a data frame with group memberships and consensus tree.

Author(s)

F. de Mendiburu

See Also

hclust, consensus, hgroups

Examples

library(agricolae)
data(pamCIP)
# only code
rownames(pamCIP)<-substr(rownames(pamCIP),1,6)
# groups of clusters
# output<-consensus(pamCIP,nboot=100)
# hcut(output,h=0.4,group=5,main="Group 5")
# 
# hcut(output,h=0.4,group=8,type="t",edgePar=list(lty=1:2,col=2:1),main="group 8"
# ,col.text="blue",cex.text=1)

[Package agricolae version 1.3-7 Index]