Data of potato, Heterosis

Determination of heterosis, general combining ability (GCA) and specific combining ability in tuber dry matter, reducing sugars and agronomic characteristics in TPS families.

The study was conducted in 3 environments, La Molina-PERU to 240 masl. during autumn-winter and spring, and in Huancayo-PERU 3180 masl., during summer. The experimental material consisted of 24 families half brother in the form of tubers derived from TPS, obtained crossing between 8 female and 3 male parents. Design used was randomized complete block with three repetitions. The experimental unit was 30 plants in two rows at a distance of 30cm between plants and 90 cm between rows. Variables evaluated were Yield, Tubers number, Dry matter and content and reducing sugars. The analysis was conducted line x tester. The control variety was Desiree.

data(heterosis)

Format

A data.frame with 216 observations on the following 11 variables.

@details

Place 1 = La Molina, 2 = Huancayo

Replication Replication

Treatment Treatment

Factor a factor with levels Control, progenie, progenitor, and testigo

Female a factor with levels Achirana, LT-8, MF-I, MF-II, Serrana, TPS-2, TPS-25, and TPS-7

Male a factor with levels TPS-13, TPS-67, and TS-15

v1 Yield (Kg/plant)

v2 Reducing sugars (scale):1 low and 5=High

v3 Tuber dry matter (percentage)

v4 Tuber number/plant

v5 Average tuber weight (g)

Source

International Potato Center(CIP). Lima-Peru. Data Kindly provided by of Rolando Cabello.

References

Thesis "Heterosis, habilidad combinatoria general y especifica para materia seca, azucares reductores y caracteres agronomicos en familias de tuberculos provenientes de semilla sexual de papa. Magister Scientiae Rodolfo Valdivia Lorente. Universidad Nacional Agraria La molina-Lima Peru, Escuela de Post Grado, Mejoramiento genetico de plantas, 2004". Poster: Congreso de la Sociedad Peruana de Genetica - Peru, 2008.

Examples

library(agricolae)
data(heterosis)
str(heterosis)

#> 'data.frame':	324 obs. of  11 variables:
#>  $ Place      : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ Replication: num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ Treatment  : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ Factor     : Factor w/ 3 levels "Control","progenie",..: 2 2 2 2 2 2 2 2 2 2 ...
#>  $ Female     : Factor w/ 8 levels "Achirana","LT-8",..: 2 2 2 6 6 6 7 7 7 8 ...
#>  $ Male       : Factor w/ 3 levels "TPS-13","TPS-67",..: 3 1 2 3 1 2 3 1 2 3 ...
#>  $ v1         : num  0.948 1.052 1.05 1.058 1.123 ...
#>  $ v2         : num  1.65 2.2 1.88 2 2.45 2.63 2.75 3 2.51 1.93 ...
#>  $ v3         : num  17.2 17.8 15.6 16 16.5 ...
#>  $ v4         : num  9.93 12.45 9.3 12.77 14.13 ...
#>  $ v5         : num  102.6 107.4 120.5 83.8 90.4 ...

site1<-subset(heterosis,heterosis[,1]==1)
site2<-subset(heterosis,heterosis[,1]==2)
site3<-subset(heterosis,heterosis[,1]==3)
model1<-with(site1,lineXtester(Replication, Female, Male, v1))

#> 
#> ANALYSIS LINE x TESTER:  v1 
#> 
#> ANOVA with parents and crosses 
#> ==============================
#>                      Df    Sum Sq    Mean Sq F value Pr(>F)
#> Replications          2 0.1022761 0.05113806   4.215 0.0187
#> Treatments           35 3.1298374 0.08942393   7.371 0.0000
#> Parents              11 0.8753481 0.07957710   6.559 0.0000
#> Parents vs. Crosses   1 1.2772782 1.27727824 105.285 0.0000
#> Crosses              23 0.9772111 0.04248744   3.502 0.0000
#> Error                70 0.8492172 0.01213167               
#> Total               107 4.0813308                          
#> 
#> ANOVA for line X tester analysis 
#> ================================
#>                 Df    Sum Sq    Mean Sq F value Pr(>F)
#> Lines            7 0.3005624 0.04293749   1.162 0.3824
#> Testers          2 0.1594550 0.07972751   2.158 0.1524
#> Lines X Testers 14 0.5171936 0.03694240   3.045 0.0010
#> Error           70 0.8492172 0.01213167               
#> 
#> ANOVA for line X tester analysis including parents 
#> ==================================================
#>                      Df    Sum Sq    Mean Sq F value Pr(>F)
#> Replications          2 0.1022761 0.05113806   4.215 0.0187
#> Treatments           35 3.1298374 0.08942393   7.371 0.0000
#> Parents              11 0.8753481 0.07957710   6.559 0.0000
#> Parents vs. Crosses   1 1.2772782 1.27727824 105.285 0.0000
#> Crosses              23 0.9772111 0.04248744   3.502 0.0000
#> Lines                 7 0.3005624 0.04293749   1.162 0.3824
#> Testers               2 0.1594550 0.07972751   2.158 0.1524
#> Lines X Testers      14 0.5171936 0.03694240   3.045 0.0010
#> Error                70 0.8492172 0.01213167               
#> Total               107 4.0813308                          
#> 
#> GCA Effects: 
#> ===========
#> Lines Effects:
#> Achirana     LT-8     MF-I    MF-II  Serrana    TPS-2   TPS-25    TPS-7 
#>   -0.079    0.069    0.046    0.021    0.075    0.016   -0.112   -0.037 
#> 
#> Testers Effects:
#> TPS-13 TPS-67  TS-15 
#> -0.003 -0.056  0.059 
#> 
#> SCA Effects: 
#> ===========
#>           Testers
#> Lines      TPS-13 TPS-67  TS-15
#>   Achirana -0.047  0.075 -0.028
#>   LT-8     -0.026  0.211 -0.185
#>   MF-I      0.071 -0.092  0.021
#>   MF-II     0.010 -0.015  0.005
#>   Serrana  -0.090 -0.035  0.125
#>   TPS-2     0.071  0.000 -0.071
#>   TPS-25   -0.019 -0.002  0.021
#>   TPS-7     0.030 -0.143  0.113
#> 
#> Standard Errors for Combining Ability Effects: 
#> =============================================
#> S.E. (gca for line)   : 0.03671463
#> S.E. (gca for tester) : 0.02248303
#> S.E. (sca effect)     : 0.0635916
#> S.E. (gi - gj)line    : 0.05192232
#> S.E. (gi - gj)tester  : 0.0317958
#> S.E. (sij - skl)tester: 0.0899321 
#> 
#> Genetic Components: 
#> ==================
#> Cov H.S. (line)   : 0.000666121
#> Cov H.S. (tester) : 0.001782713
#> Cov H.S. (average): 0.0001276635
#> Cov F.S. (average): 0.01369027
#> F = 0, Adittive genetic variance: 0.000510654
#> F = 1, Adittive genetic variance: 0.000255327
#> F = 0, Variance due to Dominance: 0.01654049
#> F = 1, Variance due to Dominance: 0.008270243 
#> 
#> Proportional contribution of lines, testers 
#>  and their interactions to total variance 
#> ===========================================
#> Contributions of lines  : 30.75717
#> Contributions of testers: 16.31736
#> Contributions of lxt    : 52.92548 

DFe <- df.residual(model1)
CMe <- deviance(model1)/DFe
test1 <- with(site1,HSD.test(v1, Factor,DFe,CMe))
test2 <- with(site1,HSD.test(v1, Treatment,DFe,CMe))
model22<-with(site2,lineXtester(Replication, Female, Male, v3))

#> 
#> ANALYSIS LINE x TESTER:  v3 
#> 
#> ANOVA with parents and crosses 
#> ==============================
#>                      Df      Sum Sq    Mean Sq F value Pr(>F)
#> Replications          2   7.2183130  3.6091565   3.519 0.0350
#> Treatments           35 179.0269741  5.1150564   4.987 0.0000
#> Parents              11 114.1375417 10.3761402  10.117 0.0000
#> Parents vs. Crosses   1   0.6154671  0.6154671   0.600 0.4412
#> Crosses              23  64.2739653  2.7945202   2.725 0.0007
#> Error                70  71.7964204  1.0256631               
#> Total               107 258.0417074                          
#> 
#> ANOVA for line X tester analysis 
#> ================================
#>                 Df   Sum Sq   Mean Sq F value Pr(>F)
#> Lines            7 14.20819  2.029741   1.085 0.4224
#> Testers          2 23.88387 11.941935   6.386 0.0107
#> Lines X Testers 14 26.18191  1.870136   1.823 0.0520
#> Error           70 71.79642  1.025663               
#> 
#> ANOVA for line X tester analysis including parents 
#> ==================================================
#>                      Df      Sum Sq    Mean Sq F value Pr(>F)
#> Replications          2   7.2183130  3.6091565   3.519 0.0350
#> Treatments           35 179.0269741  5.1150564   4.987 0.0000
#> Parents              11 114.1375417 10.3761402  10.117 0.0000
#> Parents vs. Crosses   1   0.6154671  0.6154671   0.600 0.4412
#> Crosses              23  64.2739653  2.7945202   2.725 0.0007
#> Lines                 7  14.2081875  2.0297411   1.085 0.4224
#> Testers               2  23.8838694 11.9419347   6.386 0.0107
#> Lines X Testers      14  26.1819083  1.8701363   1.823 0.0520
#> Error                70  71.7964204  1.0256631               
#> Total               107 258.0417074                          
#> 
#> GCA Effects: 
#> ===========
#> Lines Effects:
#> Achirana     LT-8     MF-I    MF-II  Serrana    TPS-2   TPS-25    TPS-7 
#>    0.752    0.050   -0.537    0.535   -0.082    0.074   -0.162   -0.630 
#> 
#> Testers Effects:
#> TPS-13 TPS-67  TS-15 
#>  0.599  0.178 -0.777 
#> 
#> SCA Effects: 
#> ===========
#>           Testers
#> Lines      TPS-13 TPS-67  TS-15
#>   Achirana  0.390 -0.179 -0.210
#>   LT-8      0.412 -1.857  1.445
#>   MF-I      0.398 -0.127 -0.272
#>   MF-II     0.440  0.041 -0.480
#>   Serrana  -0.403  0.589 -0.186
#>   TPS-2    -0.439  0.239  0.201
#>   TPS-25   -0.379  0.669 -0.289
#>   TPS-7    -0.418  0.626 -0.208
#> 
#> Standard Errors for Combining Ability Effects: 
#> =============================================
#> S.E. (gca for line)   : 0.3375834
#> S.E. (gca for tester) : 0.2067268
#> S.E. (sca effect)     : 0.5847117
#> S.E. (gi - gj)line    : 0.4774151
#> S.E. (gi - gj)tester  : 0.2923558
#> S.E. (sij - skl)tester: 0.8269071 
#> 
#> Genetic Components: 
#> ==================
#> Cov H.S. (line)   : 0.01773386
#> Cov H.S. (tester) : 0.4196583
#> Cov H.S. (average): 0.02128211
#> Cov F.S. (average): 1.418314
#> F = 0, Adittive genetic variance: 0.08512845
#> F = 1, Adittive genetic variance: 0.04256422
#> F = 0, Variance due to Dominance: 0.5629821
#> F = 1, Variance due to Dominance: 0.2814911 
#> 
#> Proportional contribution of lines, testers 
#>  and their interactions to total variance 
#> ===========================================
#> Contributions of lines  : 22.10567
#> Contributions of testers: 37.15948
#> Contributions of lxt    : 40.73486 

model3<-with(site3,lineXtester(Replication, Female, Male, v4))

#> 
#> ANALYSIS LINE x TESTER:  v4 
#> 
#> ANOVA with parents and crosses 
#> ==============================
#>                      Df    Sum Sq     Mean Sq F value Pr(>F)
#> Replications          2 0.0372615 0.018630750   4.463 0.0150
#> Treatments           35 0.8478920 0.024225486   5.803 0.0000
#> Parents              11 0.3377114 0.030701038   7.354 0.0000
#> Parents vs. Crosses   1 0.2527654 0.252765375  60.550 0.0000
#> Crosses              23 0.2574152 0.011191966   2.681 0.0008
#> Error                70 0.2922125 0.004174464               
#> Total               107 1.1773660                           
#> 
#> ANOVA for line X tester analysis 
#> ================================
#>                 Df     Sum Sq     Mean Sq F value Pr(>F)
#> Lines            7 0.13861943 0.019802776   4.510 0.0080
#> Testers          2 0.05732808 0.028664042   6.529 0.0099
#> Lines X Testers 14 0.06146769 0.004390550   1.052 0.4150
#> Error           70 0.29221250 0.004174464               
#> 
#> ANOVA for line X tester analysis including parents 
#> ==================================================
#>                      Df     Sum Sq     Mean Sq F value Pr(>F)
#> Replications          2 0.03726150 0.018630750   4.463 0.0150
#> Treatments           35 0.84789200 0.024225486   5.803 0.0000
#> Parents              11 0.33771142 0.030701038   7.354 0.0000
#> Parents vs. Crosses   1 0.25276537 0.252765375  60.550 0.0000
#> Crosses              23 0.25741521 0.011191966   2.681 0.0008
#> Lines                 7 0.13861943 0.019802776   4.510 0.0080
#> Testers               2 0.05732808 0.028664042   6.529 0.0099
#> Lines X Testers      14 0.06146769 0.004390550   1.052 0.4150
#> Error                70 0.29221250 0.004174464               
#> Total               107 1.17736600                           
#> 
#> GCA Effects: 
#> ===========
#> Lines Effects:
#> Achirana     LT-8     MF-I    MF-II  Serrana    TPS-2   TPS-25    TPS-7 
#>    0.040   -0.058   -0.027   -0.025    0.083   -0.004   -0.038    0.028 
#> 
#> Testers Effects:
#> TPS-13 TPS-67  TS-15 
#> -0.004  0.036 -0.032 
#> 
#> SCA Effects: 
#> ===========
#>           Testers
#> Lines      TPS-13 TPS-67  TS-15
#>   Achirana -0.027  0.009  0.018
#>   LT-8     -0.010 -0.015  0.025
#>   MF-I      0.030  0.008 -0.037
#>   MF-II    -0.036  0.031  0.005
#>   Serrana   0.031 -0.037  0.006
#>   TPS-2    -0.016 -0.041  0.057
#>   TPS-25   -0.011  0.026 -0.016
#>   TPS-7     0.039  0.019 -0.058
#> 
#> Standard Errors for Combining Ability Effects: 
#> =============================================
#> S.E. (gca for line)   : 0.0215367
#> S.E. (gca for tester) : 0.01318848
#> S.E. (sca effect)     : 0.03730266
#> S.E. (gi - gj)line    : 0.03045749
#> S.E. (gi - gj)tester  : 0.01865133
#> S.E. (sij - skl)tester: 0.05275392 
#> 
#> Genetic Components: 
#> ==================
#> Cov H.S. (line)   : 0.00171247
#> Cov H.S. (tester) : 0.001011396
#> Cov H.S. (average): 0.0001565892
#> Cov F.S. (average): 0.004481553
#> F = 0, Adittive genetic variance: 0.0006263566
#> F = 1, Adittive genetic variance: 0.0003131783
#> F = 0, Variance due to Dominance: 0.0001440569
#> F = 1, Variance due to Dominance: 7.202844e-05 
#> 
#> Proportional contribution of lines, testers 
#>  and their interactions to total variance 
#> ===========================================
#> Contributions of lines  : 53.85052
#> Contributions of testers: 22.27067
#> Contributions of lxt    : 23.87881