Data: Rainfall thresholds as support for timing fungicide applications in the control of potato late blight in Peru

Timing fungicide sprays based on accumulated rainfall thresholds can be a successful component of integrated management packages that include cultivars with moderate or high levels of resistance to late blight. The simplicity of measuring accumulated rainfall means that the technology can potentially be used by resource-poor farmers in developing countries.

data(huasahuasi)

Format

The format is: List of 2 ( AUDPC, YIELD )

block

a factor with levels I II III

trt

a factor with levels 40mm 7-days Non-application

clon

a factor with levels C386209.10 C387164.4 Cruza148 Musuq Yungay

y1da

a numeric vector, Kgr./plot

y2da

a numeric vector, Kgr./plot

y3ra

a numeric vector, Kgr./plot

d44

a numeric vector, 44 days

d51

a numeric vector, 51 days

d100

a numeric vector, 100 days

Details

The experimental unit was formed by 4 furrows of 1.8 m of length, with distance between furrows from 0.90 m and between plants of 0.30 m. In each furrow was installed 5 plants. The experiment had 3 repetitions. From the beginning of the experiment were fulfilled the following treatments Thresholds 40 mm: Apply the fungicide when 40 precipitation mm accumulates. The minimum interval between applications will be of 7 days. Schedule 7 days: The applications should be carried out every 7 days calendar. Without application: No fungicide application will be made. The evaluation of the severity of the late blight in each treatment started to emergency 80 percentage and then evaluations were made every 7 days until being observed a physiological maturation of the crop.

Source

Experimental field, 2003. Data Kindly provided by Wilmer Perez.

References

International Potato Center. CIP - Lima Peru.

Examples

library(agricolae)
data(huasahuasi)
names(huasahuasi)

#> [1] "AUDPC" "YIELD"

str(huasahuasi$AUDPC)

#> 'data.frame':	450 obs. of  13 variables:
#>  $ block: Factor w/ 3 levels "I","II","III": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ trt  : Factor w/ 3 levels "40mm","7-days",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ clon : Factor w/ 5 levels "C386209.10","C387164.4",..: 2 2 2 2 2 2 2 2 2 2 ...
#>  $ plant: num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ d44  : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ d51  : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ d58  : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ d65  : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ d72  : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ d79  : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ d86  : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ d93  : num  0 0 0 1 1 0 0 0 0 0 ...
#>  $ d100 : num  0 0 0 1 1 0 0 0 0 0 ...

str(huasahuasi$YIELD)

#> 'data.frame':	45 obs. of  6 variables:
#>  $ block: Factor w/ 3 levels "I","II","III": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ trt  : Factor w/ 3 levels "40mm","7-days",..: 1 1 1 1 1 2 2 2 2 2 ...
#>  $ clon : Factor w/ 5 levels "C386209.10","C387164.4",..: 2 1 3 5 4 2 1 3 5 4 ...
#>  $ y1da : num  22.8 10.8 1.6 10.9 3.5 ...
#>  $ y2da : num  5.5 8 11.7 9.9 5.1 7 12.5 9.15 8.8 4.6 ...
#>  $ y3da : num  4.14 6.1 4.15 5.2 2.4 7.6 3.6 4.9 5.9 3.5 ...