index.bio {agricolae}	R Documentation

Biodiversity Index

Description

Scientists use a formula called the biodiversity index to describe the amount of species diversity in a given area.

Usage

index.bio(data, method = c("Margalef", "Simpson.Dom", "Simpson.Div", 
"Berger.Parker", "McIntosh", "Shannon"), level=95, nboot=100, console=TRUE)

Arguments

data
number of specimens

method
Describe method bio-diversity

level
Significant level

nboot
size bootstrap

console
output console TRUE

Details

method bio-diversity. "Margalef" "Simpson.Dom" "Simpson.Div" "Berger.Parker" "McIntosh" "Shannon"

Value

Index and confidence intervals.

Author(s)

Felipe de Mendiburu

References

Magurran, A.E. (1988) Ecological diversity and its measurement. Princeton University Press Efron, B., Tibshirani, R. (1993) An Introduction to the Boostrap. Chapman and Hall/CRC

Examples

library(agricolae)
data(paracsho)
# date 22-06-05 and treatment CON = application with insecticide
specimens <- paracsho[1:10,6]
output1 <- index.bio(specimens,method="Simpson.Div",level=95,nboot=100)
output2 <- index.bio(specimens,method="Shannon",level=95,nboot=100)
rbind(output1, output2)

[Package agricolae version 1.3-7 Index]