join.freq {agricolae}	R Documentation

Join class for histogram

Description

In many situations it is required to join classes because of the low .frequency in the intervals. In this process, it is required to join the intervals and ad the .frequencies of them.

Usage

join.freq(histogram, join)

Arguments

histogram
Class graph.freq

join
vector

Value

New histogram with union of classes.

Author(s)

Felipe de Mendiburu

See Also

polygon.freq, table.freq, stat.freq, inter.freq, sturges.freq, graph.freq, ogive.freq, normal.freq

Examples

library(agricolae)
data(natives)
# histogram
h1<-graph.freq(natives$size,plot=FALSE)
round(table.freq(h1),4)
# Join classes  9, 10,11 and 12 with little frequency.
h2<-join.freq(h1,9:12)
# new table
plot(h2,col="bisque",xlab="Size")
round(summary(h2),4)

[Package agricolae version 1.3-7 Index]