lineXtester {agricolae}	R Documentation

Line x Tester Analysis

Description

It makes the Line x Tester Genetic Analysis. It also estimates the general and specific combinatory ability effects and the line and tester genetic contribution.

Usage

lineXtester(replications, lines, testers, y)

Arguments

replications
Replications

lines
Lines

testers
Testers

y
Variable, response

Details

ANOVA with parents and crosses
ANOVA for line X tester analysis
ANOVA for line X tester analysis including parents
GCA Effects: Lines Effects, Testers Effects and SCA Effects.
Standard Errors for Combining Ability Effects.
Genetic Components.
...
Proportional contribution of lines, testers and their interactions to total variance

Value

return anova(formula = Y ~ Replications + Treatments).
where the Treatments contains parents, crosses and crosses vs Parents.
The crosses contains Lines, Testers and its interaction .

Author(s)

Felipe de Mendiburu

References

Biometrical Methods in Quantitative Genetic Analysis, Singh, Chaudhary. 1979. Hierarchial and factorial mating designs for quantitative genetic analysis in tetrasomic potato. R. Ortis; A.Golmirzaie. Theor Appl Genet (2002) 104:675-679

See Also

AMMI

Examples

# see structure line by testers
library(agricolae)
# example 1
data(heterosis)
site1<-subset(heterosis,heterosis[,1]==1)
output1<-with(site1,lineXtester(Replication, Female, Male, v2))
# example 2
data(LxT)
str(LxT)
output2<-with(LxT,lineXtester(replication, line, tester, yield))

[Package agricolae version 1.3-7 Index]