normal.freq {agricolae}	R Documentation

Normal curve on the histogram

Description

A normal distribution graph elaborated from the histogram previously constructed. The average and variance are obtained from the data grouped in the histogram.

Usage

normal.freq(histogram, frequency=1, ...)

Arguments

histogram
object constructed by the function hist

frequency
1=counts, 2=relative, 3=density

...
Other parameters of the function hist

Author(s)

Felipe de Mendiburu

See Also

polygon.freq, table.freq, stat.freq, inter.freq, sturges.freq, join.freq, ogive.freq, graph.freq

Examples

library(agricolae)
data(growth)
#startgraph
h1<-with(growth,hist(height,col="green",xlim=c(6,14)))
normal.freq(h1,col="blue")
#endgraph
#startgraph
h2<-with(growth,graph.freq(height,col="yellow",xlim=c(6,14),frequency=2))
normal.freq(h2,frequency=2)
#endgraph

[Package agricolae version 1.3-7 Index]