polygon.freq {agricolae}	R Documentation

The polygon of frequency on the histogram

Description

The polygon is constructed single or on a histogram. It is necessary to execute the function previously hist.

Usage

polygon.freq(histogram, frequency=1, ...)

Arguments

histogram
Object constructed by the function hist

frequency
numeric, counts(1), relative(2) and density(3)

...
Other parameters of the function hist

Author(s)

Felipe de Mendiburu Delgado

See Also

polygon.freq, table.freq, stat.freq, inter.freq, sturges.freq, join.freq, graph.freq, normal.freq

Examples

library(agricolae)
data(growth)
#startgraph
h1<-with(growth,hist(height,border=FALSE,xlim=c(6,14)))
polygon.freq(h1,frequency=1,col="red")
#endgraph
#startgraph
h2<-with(growth,graph.freq(height,frequency=2,col="yellow",xlim=c(6,14)))
polygon.freq(h2,frequency=2,col="red")
#endgraph

[Package agricolae version 1.3-7 Index]