Data of assessment of the population in the soil R.solanacearum

The assessment of the population of R.solanacearum on the floor took place after 48 hours of infestation, during days 15, 29, 43, 58, and 133 days after the infestation soil. More information on soil data(soil).

Logarithm average counts of colonies on plates containing half of M-SMSA 3 repetitions (3 plates by repetition) incubated at 30 degrees centigrade for 48 hours. log(1+UFC/g soil)

data(ralstonia)

Format

A data.frame with 13 observations on the following 8 variables.

#' @details

place a factor with levels Chmar, Chz, Cnt1, Cnt2, Cnt3, Hco1, Hco2, Hco3, Hyo1, Hyo2, Namora, SR1, and SR2

Day2 Day2

Day15 Day15

Day29 Day29

Day43 Day43

Day58 Day58

Day73 Day73

Day133 Day133

Source

Experimental field, 2004. Data Kindly provided by Dr. Sylvie Priou, Liliam Gutarra and Pedro Aley.

References

International Potato Center. CIP - Lima Peru.

Examples

library(agricolae)
data(ralstonia)
str(ralstonia)

#> 'data.frame':	13 obs. of  8 variables:
#>  $ place : Factor w/ 13 levels "Chmar","Chz",..: 11 9 10 12 13 3 4 5 2 1 ...
#>  $ Day2  : num  7.51 7.4 7.4 7.42 7.33 ...
#>  $ Day15 : num  7.55 6.72 6.12 6.86 7.41 ...
#>  $ Day29 : num  6.54 3.39 2.19 6.8 7.22 ...
#>  $ Day43 : num  6.39 1.66 0 6.51 7 ...
#>  $ Day58 : num  6.21 0 0 6.52 6.94 ...
#>  $ Day73 : num  5.86 0 0 6.25 6.78 ...
#>  $ Day133: num  4.84 0 0 5.45 6.44 ...