scheffe.test {agricolae}	R Documentation

Multiple comparisons, scheffe

Description

Scheffe 1959, method is very general in that all possible contrasts can be tested for significance and confidence intervals can be constructed for the corresponding linear. The test is conservative.

Usage

scheffe.test(y, trt, DFerror, MSerror, Fc, alpha = 0.05, group=TRUE, main = NULL,
console=FALSE )

Arguments

y
model(aov or lm) or answer of the experimental unit

trt
Constant( only y=model) or vector treatment applied to each experimental unit

DFerror
Degrees of freedom

MSerror
Mean Square Error

Fc
F Value

alpha
Significant level

group
TRUE or FALSE

main
Title

console
logical, print output

Details

It is necessary first makes a analysis of variance.

if y = model, then to apply the instruction:
scheffe.test (model, "trt", alpha = 0.05, group = TRUE, main = NULL, console = FALSE)
where the model class is aov or lm.

Value

statistics
Statistics of the model

parameters
Design parameters

means
Statistical summary of the study variable

comparison
Comparison between treatments

groups
Formation of treatment groups

Author(s)

Felipe de Mendiburu

References

Robert O. Kuehl. 2nd ed. Design of experiments. Duxbury, copyright 2000.
Steel, R.; Torri,J; Dickey, D.(1997) Principles and Procedures of Statistics A Biometrical Approach. pp189

See Also

BIB.test, DAU.test, duncan.test, durbin.test, friedman, HSD.test, kruskal, LSD.test, Median.test, PBIB.test, REGW.test, SNK.test, waerden.test, waller.test, plot.group

Examples

library(agricolae)
data(sweetpotato)
model<-aov(yield~virus, data=sweetpotato)
comparison <- scheffe.test(model,"virus", group=TRUE,console=TRUE,
main="Yield of sweetpotato\nDealt with different virus")
# Old version scheffe.test()
df<-df.residual(model)
MSerror<-deviance(model)/df
Fc<-anova(model)["virus",4]
out <- with(sweetpotato,scheffe.test(yield, virus, df, MSerror, Fc))
print(out)

[Package agricolae version 1.3-7 Index]