Data of soil analysis for 13 localities

We analyzed the physical and chemical properties of different soils, as full characterization of soil and special analysis of micro-elements. These analyses were conducted in the laboratory analysis of soils, plants, water and fertilizers in the La Molina National Agrarian University (UNALM). To which the different soil samples were dried to the environment, screened (mesh 0.5xo, 5 mm) and sterilized by steam 4 to 5 hours with a Lindinger Steam aerator SA150 and SA700, with the possible aim of eliminating bacteria saprophytic or antagonists to prevent the growth of bacteria (R.solanacearum).

Cnt1= Canete, Cnt2=Valle Dulce(Canete), Cnt3=Valle Grande(Canete), Chz=Obraje-Carhuaz(Ancash), Chmar=Chucmar-Chota(Huanuco, Hco1= Mayobamba-Chinchao(Huanuco), Hco2=Nueva Independencia-Chinchao(Huanuco), Hco3=San Marcos-Umari(Huanuco), Hyo1=La Victoria-Huancayo(Junin), Hyo1=El Tambo-Huancayo(Junin), Namora=Namora(Cajamarca), SR1= El Milagro-San Ramon(Junin), Sr2=La Chinchana-San Ramon(Junin).

data(soil)

Format

A data.frame with 13 observations on the following 23 variables.

@details

place a factor with levels Chmar, Chz, Cnt1, Cnt2, Cnt3, Hco1, Hco2, Hco3, Hyo1, Hyo2, Namora, SR1, and SR2

pH pH

EC electrical conductivity

CaCO3 CaCO3

MO MO

CIC CIC

P P

K K

sand sand

slime slime

clay clay

Ca Ca

Mg Mg

K2 K2

Na Na

Al_H Al_H

K_Mg K_Mg

Ca_Mg Ca_Mg

B B

Cu Cu

Fe Fe

Mn Mn

Zn Zn

Source

Experimental field, 2004. Data Kindly provided by Dr. Sylvie Priou, Liliam Gutarra and Pedro Aley.

References

International Potato Center - Lima, PERU.

Examples

library(agricolae)
data(soil)
str(soil)

#> 'data.frame':	13 obs. of  23 variables:
#>  $ place: Factor w/ 13 levels "Chmar","Chz",..: 11 9 10 12 13 3 4 5 2 1 ...
#>  $ pH   : num  3.8 7.2 8.4 4.8 4.4 7.6 8.3 8.4 6.1 4.4 ...
#>  $ EC   : num  0.72 1 0.32 0.83 0.49 1.09 3.67 7.17 0.15 0.18 ...
#>  $ CaCO3: num  0 1.4 9.6 0 0 0.9 7.6 2.7 0 0 ...
#>  $ MO   : num  1.7 2.6 2 1.9 2 1.3 1.7 1.1 1.6 3.8 ...
#>  $ CIC  : num  5.76 20 18.88 17.28 6.4 ...
#>  $ P    : num  8.1 17.5 17.6 15.8 7.2 8.9 20 50 8.5 44.1 ...
#>  $ K    : num  28 166 124 86 30 350 537 414 84 240 ...
#>  $ sand : num  80 26 44 58 78 48 36 50 52 68 ...
#>  $ slime: num  16 42 32 30 14 38 36 28 28 26 ...
#>  $ clay : num  4 32 24 12 8 14 28 22 20 6 ...
#>  $ Ca   : num  0.74 17.21 14.91 12.49 0.76 ...
#>  $ Mg   : num  0.15 2.06 3.37 1.9 0.13 1.79 4.44 2.67 1.54 0.56 ...
#>  $ K2   : num  0.1 0.47 0.36 0.26 0.09 0.88 1.42 0.76 0.25 0.56 ...
#>  $ Na   : num  0.27 0.26 0.24 0.32 0.26 0.37 0.62 0.92 0.29 0.27 ...
#>  $ Al_H : num  1.2 0 0 0.4 0.9 0 0 0 0 2.6 ...
#>  $ K_Mg : num  0.667 0.228 0.107 0.137 0.692 ...
#>  $ Ca_Mg: num  4.93 8.35 4.42 6.57 5.85 ...
#>  $ B    : num  0.9 0.2 0.4 0.5 0.4 2.5 3.1 4.9 0.4 1.1 ...
#>  $ Cu   : num  0.4 12.1 2.5 3 0.6 5.1 1.7 1.5 1.9 2.3 ...
#>  $ Fe   : num  295.4 12.9 23.2 352.4 334.4 ...
#>  $ Mn   : num  1.3 0.5 7.9 5.1 1.8 2.1 2.1 4.6 20.6 7.6 ...
#>  $ Zn   : num  2.1 62.2 6.7 7 1.7 3.1 8.3 6.3 3.2 4.5 ...