sp.plot {agricolae}	R Documentation

Splip-Plot analysis

Description

The variance analysis of a split plot design is divided into two parts: the plot-factor analysis and the sub-plot factor analysis.

Usage

sp.plot(block, pplot, splot, Y)

Arguments

block
replications

pplot
main-plot Factor

splot
sub-plot Factor

Y
Variable, response

Details

The split-plot design is specifically suited for a two-factor experiment on of the factors is assigned to main plot (main-plot factor), the second factor, called the subplot factor, is assigned into subplots. The model is mixed, the blocks are random and the study factors are fixed applied according to the design.

Value

ANOVA: Splip plot analysis

Author(s)

Felipe de Mendiburu

References

Statistical procedures for agricultural research. Kwanchai A. Gomez, Arturo A. Gomez. Second Edition. 1984.

See Also

ssp.plot, strip.plot, design.split, design.strip

Examples

library(agricolae)
data(plots)
model<-with(plots,sp.plot(block,A,B,yield))
# with aov
plots[,1]<-as.factor(plots[,1])
AOV <- aov(yield ~ block + A*B + Error(block/A),data=plots)
summary(AOV)

[Package agricolae version 1.3-7 Index]