stability.nonpar {agricolae}	R Documentation

Nonparametric stability analysis

Description

A method based on the statistical ranges of the study variable per environment for the stability analysis.

Usage

stability.nonpar(data, variable = NULL, ranking = FALSE, console=FALSE)

Arguments

data
First column the genotypes following environment

variable
Name of variable

ranking
logical, print ranking

console
logical, print output

Value

ranking
data frame

statistics
Statistical analysis chi square test

Author(s)

Felipe de Mendiburu

References

Haynes K G, Lambert D H, Christ B J, Weingartner D P, Douches D S, Backlund J E, Fry W and Stevenson W. 1998. Phenotypic stability of resistance to late blight in potato clones evaluated at eight sites in the United States American Journal Potato Research 75, pag 211-217.

See Also

stability.par

Examples

library(agricolae)
data(haynes)
stability.nonpar(haynes,"AUDPC",ranking=TRUE,console=TRUE)
# Example 2
data(CIC)
data1<-CIC$comas[,c(1,6,7,17,18)]
data2<-CIC$oxapampa[,c(1,6,7,19,20)]
cic <- rbind(data1,data2)

means <- by(cic[,5], cic[,c(2,1)], function(x) mean(x,na.rm=TRUE))
means <-as.data.frame(means[,])
cic.mean<-data.frame(genotype=row.names(means),means)
cic.mean<-delete.na(cic.mean,"greater")
out<-stability.nonpar(cic.mean)
out$ranking
out$statistics

[Package agricolae version 1.3-7 Index]