table.freq {agricolae}	R Documentation

frequency Table of a Histogram

Description

It finds the absolute, relative and accumulated frequencies with the class intervals defined from a previously calculated histogram by the "hist" of R function.

Usage

table.freq(object)

Arguments

object
Object by function graph.freq()

Value

Frequency table.

Lower
Lower limit class

Upper
Upper limit class

Main
class point

Frequency
Frequency

Percentage
Percentage frequency

CF
Cumulative frequency

CPF
Cumulative Percentage frequency

Author(s)

Felipe de Mendiburu

See Also

polygon.freq, stat.freq, graph.freq, inter.freq, sturges.freq, join.freq, ogive.freq, normal.freq

Examples

library(agricolae)
data(growth)
h2<-with(growth,graph.freq(height,plot=FALSE))
print(table.freq(h2),row.names=FALSE)

[Package agricolae version 1.3-7 Index]