Computations of Bayesian t-values for multiple comparisons

A Bayes rule for the symmetric multiple comparisons problem.

waller(K, q, f, Fc)

Arguments

K
Is the loss ratio between type I and type II error

q
Numerator Degrees of freedom

f
Denominator Degrees of freedom

Fc
F ratio from an analysis of variance

Value

Waller value for the Waller and Duncan test.

Details

K-RATIO (K): value specifies the Type 1/Type 2 error seriousness ratio for the Waller-Duncan test. Reasonable values for KRATIO are 50, 100, and 500, which roughly correspond for the two-level case to ALPHA levels of 0.1, 0.05, and 0.01. By default, the procedure uses the default value of 100.

References

Waller, R. A. and Duncan, D. B. (1969). A Bayes Rule for the Symmetric Multiple Comparison Problem, Journal of the American Statistical Association 64, pages 1484-1504.

Waller, R. A. and Kemp, K. E. (1976) Computations of Bayesian t-Values for Multiple Comparisons, Journal of Statistical Computation and Simulation, 75, pages 169-172.

Principles and procedures of statistics a biometrical approach Steel & Torry & Dickey. Third Edition 1997.

See also

Examples

# Table Duncan-Waller K=100, F=1.2 pag 649 Steel & Torry
library(agricolae)
K<-100
Fc<-1.2
q<-c(8,10,12,14,16,20,40,100)
f<-c(seq(4,20,2),24,30,40,60,120)
n<-length(q)
m<-length(f)
W.D <-rep(0,n*m)
dim(W.D)<-c(n,m)
for (i in 1:n) {
for (j in 1:m) {
W.D[i,j]<-waller(K, q[i], f[j], Fc)
}}
W.D<-round(W.D,2)
dimnames(W.D)<-list(q,f)
print(W.D)

#>        4    6    8   10   12   14   16   18   20   24   30   40   60  120
#> 8   2.85 2.91 2.94 2.96 2.97 2.98 2.99 2.99 2.99 3.00 3.00 3.00 3.00 3.00
#> 10  2.85 2.93 2.98 3.01 3.04 3.05 3.06 3.07 3.08 3.09 3.10 3.10 3.11 3.12
#> 12  2.85 2.95 3.01 3.05 3.08 3.10 3.12 3.13 3.14 3.16 3.17 3.19 3.20 3.21
#> 14  2.85 2.96 3.03 3.08 3.12 3.14 3.16 3.18 3.19 3.21 3.23 3.25 3.27 3.29
#> 16  2.85 2.97 3.05 3.11 3.15 3.18 3.20 3.22 3.24 3.26 3.29 3.31 3.34 3.36
#> 20  2.85 2.99 3.08 3.14 3.19 3.23 3.26 3.28 3.30 3.33 3.37 3.40 3.44 3.47
#> 40  2.85 3.02 3.13 3.22 3.29 3.35 3.39 3.43 3.47 3.52 3.58 3.64 3.72 3.79
#> 100 2.85 3.04 3.17 3.28 3.36 3.44 3.50 3.55 3.59 3.67 3.76 3.86 3.98 4.11