Data Yacon

The yacon (Smallanthus sonchifolius) is a plant native to the Andes, considered a traditional crop in Peru and natural source of FOS, which is a type of carbohydrate that can not be digested by the and the human body that have joined several beneficial properties in health, such as improve the absorption of calcium, reducing the level of triglycerides and cholesterol and stimulate better gastrointestinal function.

Proportion or fraction of the plant that is used (seeds, fruit, root) on dry basis. Part usable in a proportion of total mass dissected. Plant of frijol, weight = 100g and frijol = 50g then, IH = 50/100 = 0.5 or 50 percentaje. Degrees Brix is a measurement of the mass ratio of dissolved sugar to water in a liquid.

data(yacon)

Format

A data.frame with 432 observations on the following 19 variables.

Source

CIP. Experimental field, 2003, Data Kindly provided by Ivan Manrique and Carolina Tasso.

Details

locality a factor with levels, Cajamarca, Lima, Oxapampa in PERU

site site

dose a factor with levels F0, F150, and F80

entry a factor with levels AKW5075, AMM5136, AMM5150, AMM5163, ARB5125, CLLUNC118, P1385, and SAL136

replication replication

height plant height, centimeters

stalks number of stalks

wfr weight of fresh roots, grams

wff weight of fresh foliage, grams

wfk weight fresh kroner, grams

roots matter of dried roots, grams

FOS fructo-oligosaccharides, percentage

glucose glucose

fructose fructose

sucrose sucrose

brix brix

foliage foliage

dry dry

IH IH

References

International Potato Center. CIP - Lima Peru.

Examples

library(agricolae)
data(yacon)
str(yacon)

#> 'data.frame':	432 obs. of  19 variables:
#>  $ locality   : Factor w/ 3 levels "CAJ","LIM","OXA": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ site       : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ dose       : Factor w/ 3 levels "F0","F150","F80": 1 1 1 3 3 3 2 2 2 1 ...
#>  $ entry      : Factor w/ 8 levels "AKW5075 ","AMM5136",..: 7 7 7 7 7 7 7 7 7 3 ...
#>  $ replication: num  1 2 3 1 2 3 1 2 3 1 ...
#>  $ height     : num  57.9 62.1 53.7 55.8 61.5 66.5 68 52.8 60.5 70.5 ...
#>  $ stalks     : num  3.2 3 2.6 3.7 3 3.6 3.1 5.6 3.4 3 ...
#>  $ wfr        : num  6700 6450 7550 4120 7150 7520 6700 6450 7550 9420 ...
#>  $ wff        : num  3600 2500 2450 2940 4000 4450 3600 2500 2450 2520 ...
#>  $ wfk        : num  5490 3800 5360 3580 4560 5180 5490 3800 5360 4650 ...
#>  $ roots      : num  21.9 21.3 21.8 17.7 16.8 17.3 17.3 16.4 19.6 22.8 ...
#>  $ FOS        : num  60.2 60.4 54.5 58.5 44.2 50.3 49.6 42 57.5 67.6 ...
#>  $ glucose    : num  1.74 1.5 2.14 2.5 2.8 3 2.6 4.2 1.79 1.5 ...
#>  $ fructose   : num  5 4.5 9.3 8.1 10.6 13.1 9.9 14.7 7.75 1.5 ...
#>  $ sucrose    : num  26.7 28.8 27.1 25.6 27.9 ...
#>  $ brix       : num  14.8 15.7 14.3 12.4 13 13.3 13.4 12.4 15.1 15.9 ...
#>  $ foliage    : num  25.6 27.2 30 27.6 27.5 22 24.6 26.5 25.8 28.7 ...
#>  $ dry        : num  15.6 17 15.9 13.7 14.5 15.1 19.4 16.5 21.5 18.2 ...
#>  $ IH         : num  0.347 0.326 0.322 0.3 0.286 ...