zigzag {agricolae}	R Documentation

order plot in serpentine

Description

applied to designs: complete block, latin square, graeco, split plot, strip plot, lattice, alpha lattice, Augmented block, cyclic, Balanced Incomplete Block and factorial.

Usage

zigzag(outdesign)

Arguments

outdesign
output design

Value

fieldbook
Remuneration of serpentine plots.

Author(s)

Felipe de Mendiburu

See Also

design.ab, design.alpha,design.bib, design.split , design.cyclic , design.dau , design.graeco, design.lattice, design.lsd, design.rcbd, design.strip

Examples

library(agricolae)
trt<-letters[1:5]
r<-4
outdesign <- design.rcbd(trt,r,seed=9)
fieldbook <- zigzag(outdesign)

[Package agricolae version 1.3-7 Index]