Data for the analysis of carolina genetic design

Data for the analysis of carolina I, II and III genetic design

data(CIC)

Format

DC is list, 3 data.frame: carolina1(72 obs, 6 var), carolina2(300 obs, 9 var) and carolina3(64 obs, 5 var).

Source

Biometrical Methods in Quantitative Genetic Analysis, Singh, Chaudhary. 1979.

Details

Carolina1 Data for the analysis of Carolina I Genetic design. In this design F2 or any advanced generation maintained by random mating, produced from cross between two pure-lines, is taken as base population. From the population an individual is randomly selected and used as a male. A set of 4 randomly selected plans are used as females and are mated to the above male. Thus a set of 4 full-sib families are produced. This is denoted as a male group. Similarly, a large number of male groups are produced. No female is used for any second mating. four male groups (16 female groups) from a set.

Carolina2 Data for the analysis of Carolina II Genetic design. Both paternal and maternal half-sibs are produced in this design. From an F2 population, n1 males and n2 females are randomly selected and each male is crossed to each of the females. Thus n1 x n2 progenies are produced whitch are analysed in a suitably laid experiment.

Carolina3 Data for the analysis of Carolina III genetic design. The F2 population is produced by crossing two inbreds, say L1 and L2. The material for estimation of genetic parameters is produced by back crossing randomly selected F2 individuals (using as males) to each of the inbreds (used as females).

References

Biometrical Methods in Quantitative Genetic Analysis, Singh, Chaudhary. 1979.

Examples

data(DC)
names(DC)

#> [1] "carolina1" "carolina2" "carolina3"

str(DC$carolina1)

#> 'data.frame':	72 obs. of  6 variables:
#>  $ set     : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ male    : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ female  : int  1 1 1 2 2 2 1 1 1 2 ...
#>  $ progenie: int  1 2 3 1 2 3 1 2 3 1 ...
#>  $ rep     : int  1 1 1 1 1 1 2 2 2 2 ...
#>  $ yield   : num  3.6 3.4 3.1 3.8 3.6 3.2 3.5 3.7 3.6 3.1 ...

str(DC$carolina2)

#> 'data.frame':	300 obs. of  9 variables:
#>  $ Loc   : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ set   : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ rep   : int  1 2 3 1 2 3 1 2 3 1 ...
#>  $ female: int  1 1 1 2 2 2 3 3 3 4 ...
#>  $ male  : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ plrv  : num  52 52 60 36 40 ...
#>  $ yield : num  831 983 679 626 462 ...
#>  $ tuber : num  5.59 6.61 4.34 4.63 2.79 4.37 3.23 3.57 3.52 5.33 ...
#>  $ weight: num  149 149 156 135 165 ...

str(DC$carolina3)

#> 'data.frame':	64 obs. of  5 variables:
#>  $ set   : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ male  : num  1 2 3 4 1 2 3 4 1 2 ...
#>  $ female: num  1 1 1 1 2 2 2 2 1 1 ...
#>  $ rep   : num  1 1 1 1 1 1 1 1 2 2 ...
#>  $ yield : num  4.2 5.7 5.3 2.2 3.8 3.7 3.3 5.8 3.8 5 ...