Data and analysis Mother and baby trials

Mother/Baby Trials allow farmers and researchers to test best-bet technologies or new cultivars. Evaluation of advanced Clones of potato in the Valley of Rio Chillon - PERU (2004)

Format

The format is list of 2:
1. mother: data.frame: 30 obs. of 3 variables:
- block (3 levels)
- clon (10 levels)
- yield (kg.)
2. babies: data.frame: 90 obs. of 3 variables:
- farmer (9 levels)
- clon (10 levels)
- yield (kg.)

Source

Experimental field.

Details

1. Replicated researcher-managed "mother trials" with typically 10 treatments evaluated in small plots.
2. Unreplicated "baby trials" with 10 treatments evaluated in large plots.
3. The "baby trials" with a subset of the treatments in the mother trial.

References

International Potato Center. CIP - Lima Peru.

Examples

# Analisys the Mother/Baby Trial Design
library(agricolae)
data(RioChillon)
# First analysis the Mother Trial Design.
model<-aov(yield ~ block + clon, RioChillon$mother)
anova(model)

#> Analysis of Variance Table
#> 
#> Response: yield
#>           Df  Sum Sq Mean Sq F value  Pr(>F)   
#> block      2  20.909 10.4543  2.1240 0.14854   
#> clon       9 177.783 19.7537  4.0133 0.00586 **
#> Residuals 18  88.598  4.9221                   
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

cv.model(model)

#> [1] 34.64732

comparison<-with(RioChillon$mother,LSD.test(yield,clon, 18, 4.922, group=TRUE))
# Second analysis the babies Trial.
comparison<-with(RioChillon$babies,friedman(farmer,clon, yield, group=TRUE))
# Third
# The researcher makes use of data from both mother and baby trials and thereby obtains
# information on suitability of new technologies or cultivars
# for different agro-ecologies and acceptability to farmers.