Design of experiments for a factorial

It generates a design of blocks, randomize and latin square for combined n. factors uses the methods of number generation in R. The seed is by set.seed(seed, kinds).

design.ab(
  trt,
  r = NULL,
  serie = 2,
  design = c("rcbd", "crd", "lsd"),
  seed = 0,
  kinds = "Super-Duper",
  first = TRUE,
  randomization = TRUE
)

Arguments

trt
n levels factors

r
Replications or Blocks

serie
number plot, 1: 11,12; 2: 101,102; 3: 1001,1002

design
type

seed
Seed

kinds
Method for to randomize

first
TRUE or FALSE - randomize rep 1

randomization
TRUE or FALSE - randomize

Value

parameters

Design parameters

book

Fieldbook

Details

kinds <- c("Wichmann-Hill", "Marsaglia-Multicarry", "Super-Duper", "Mersenne-Twister", "Knuth-TAOCP", "user-supplied", "Knuth-TAOCP-2002", "default" )

References

Introduction to Experimental Statistics. Ching Chun Li. McGraw-Hill Book Company, INC, New. York, 1964

See also

Examples

# factorial 3 x 2 with 3 blocks
library(agricolae)
trt<-c(3,2) # factorial 3x2
outdesign <-design.ab(trt, r=3, serie=2)
book<-outdesign$book
head(book,10) # print of the field book

#>    plots block A B
#> 1    101     1 2 2
#> 2    102     1 1 2
#> 3    103     1 3 2
#> 4    104     1 2 1
#> 5    105     1 1 1
#> 6    106     1 3 1
#> 7    107     2 2 2
#> 8    108     2 3 2
#> 9    109     2 2 1
#> 10   110     2 1 2

# factorial 2 x 2 x 2 with 5 replications in completely randomized design.
trt<-c(2,2,2)
outdesign<-design.ab(trt, r=5, serie=2,design="crd")
book<-outdesign$book
print(book)

#>    plots r A B C
#> 1    101 1 1 1 2
#> 2    102 1 1 2 2
#> 3    103 1 2 1 1
#> 4    104 1 2 2 1
#> 5    105 1 2 2 2
#> 6    106 1 2 1 2
#> 7    107 2 1 1 2
#> 8    108 2 2 2 1
#> 9    109 2 2 2 2
#> 10   110 2 1 2 2
#> 11   111 1 1 1 1
#> 12   112 2 2 1 1
#> 13   113 3 2 2 2
#> 14   114 3 1 2 2
#> 15   115 3 2 1 1
#> 16   116 3 2 2 1
#> 17   117 3 1 1 2
#> 18   118 1 1 2 1
#> 19   119 2 1 2 1
#> 20   120 3 1 2 1
#> 21   121 4 1 2 2
#> 22   122 4 2 2 2
#> 23   123 2 2 1 2
#> 24   124 4 1 2 1
#> 25   125 4 2 1 1
#> 26   126 2 1 1 1
#> 27   127 5 2 1 1
#> 28   128 3 1 1 1
#> 29   129 4 2 2 1
#> 30   130 4 1 1 2
#> 31   131 5 1 2 2
#> 32   132 4 1 1 1
#> 33   133 5 2 2 1
#> 34   134 5 1 1 2
#> 35   135 5 1 1 1
#> 36   136 3 2 1 2
#> 37   137 4 2 1 2
#> 38   138 5 2 2 2
#> 39   139 5 2 1 2
#> 40   140 5 1 2 1

# factorial 3 x 3 in latin square design.
trt <-c(3,3)
outdesign<-design.ab(trt, serie=2, design="lsd")
book<-outdesign$book
print(book)

#>    plots row col A B
#> 1    101   1   1 3 3
#> 2    102   1   2 2 3
#> 3    103   1   3 1 2
#> 4    104   1   4 3 2
#> 5    105   1   5 1 1
#> 6    106   1   6 2 2
#> 7    107   1   7 1 3
#> 8    108   1   8 3 1
#> 9    109   1   9 2 1
#> 10   201   2   1 3 1
#> 11   202   2   2 2 1
#> 12   203   2   3 3 3
#> 13   204   2   4 2 3
#> 14   205   2   5 3 2
#> 15   206   2   6 1 3
#> 16   207   2   7 1 1
#> 17   208   2   8 2 2
#> 18   209   2   9 1 2
#> 19   301   3   1 1 1
#> 20   302   3   2 3 1
#> 21   303   3   3 1 3
#> 22   304   3   4 3 3
#> 23   305   3   5 1 2
#> 24   306   3   6 2 3
#> 25   307   3   7 2 1
#> 26   308   3   8 3 2
#> 27   309   3   9 2 2
#> 28   401   4   1 1 2
#> 29   402   4   2 3 2
#> 30   403   4   3 2 1
#> 31   404   4   4 1 1
#> 32   405   4   5 1 3
#> 33   406   4   6 3 1
#> 34   407   4   7 2 2
#> 35   408   4   8 3 3
#> 36   409   4   9 2 3
#> 37   501   5   1 2 1
#> 38   502   5   2 1 1
#> 39   503   5   3 2 3
#> 40   504   5   4 1 3
#> 41   505   5   5 2 2
#> 42   506   5   6 3 3
#> 43   507   5   7 3 1
#> 44   508   5   8 1 2
#> 45   509   5   9 3 2
#> 46   601   6   1 2 3
#> 47   602   6   2 1 3
#> 48   603   6   3 3 2
#> 49   604   6   4 2 2
#> 50   605   6   5 3 1
#> 51   606   6   6 1 2
#> 52   607   6   7 3 3
#> 53   608   6   8 2 1
#> 54   609   6   9 1 1
#> 55   701   7   1 1 3
#> 56   702   7   2 3 3
#> 57   703   7   3 2 2
#> 58   704   7   4 1 2
#> 59   705   7   5 2 1
#> 60   706   7   6 3 2
#> 61   707   7   7 2 3
#> 62   708   7   8 1 1
#> 63   709   7   9 3 1
#> 64   801   8   1 2 2
#> 65   802   8   2 1 2
#> 66   803   8   3 3 1
#> 67   804   8   4 2 1
#> 68   805   8   5 2 3
#> 69   806   8   6 1 1
#> 70   807   8   7 3 2
#> 71   808   8   8 1 3
#> 72   809   8   9 3 3
#> 73   901   9   1 3 2
#> 74   902   9   2 2 2
#> 75   903   9   3 1 1
#> 76   904   9   4 3 1
#> 77   905   9   5 3 3
#> 78   906   9   6 2 1
#> 79   907   9   7 1 2
#> 80   908   9   8 2 3
#> 81   909   9   9 1 3