Alpha design type (0,1)

Generates an alpha designs starting from the alpha design fixing under the series formulated by Patterson and Williams. These designs are generated by the alpha arrangements. They are similar to the lattice designs, but the tables are rectangular s by k (with s blocks and k<s columns. The number of treatments should be equal to s*k and all the experimental units r*s*k (r replications).

design.alpha(trt, k, r, serie = 2, seed = 0, kinds = "Super-Duper",randomization=TRUE)

Arguments

trt
Treatments

k
size block

r
Replications

serie
number plot, 1: 11,12; 2: 101,102; 3: 1001,1002

seed
seed

kinds
method for to randomize

randomization
TRUE or FALSE - randomize

Details

Parameters for the alpha design: I. r=2, k <= s; II. r=3, s odd, k <= s; III.r=3, s even, k <= s-1; IV. r=4, s odd but not a multiple of 3, k<=s

r= replications s=number of blocks k=size of block Number of treatment is equal to k*s

Value

parameters

Design parameters

statistics

Design statistics

sketch

Design sketch

book

Fieldbook

References

H.D. Patterson and E.R. Williams. Biometrika (1976) A new class of resolvable incomplete block designs. printed in Great Britain. Online: http://biomet.oxfordjournals.org/cgi/content/abstract/63/1/83

See also

Examples

library(agricolae)
#Example one
trt<-1:30
t <- length(trt)
# size block k
k<-3
# Blocks s
s<-t/k
# replications r
r <- 2
outdesign<- design.alpha(trt,k,r,serie=2)

#> 
#> Alpha Design (0,1) - Serie  I 
#> 
#> Parameters Alpha Design
#> =======================
#> Treatmeans : 30
#> Block size : 3
#> Blocks     : 10
#> Replication: 2 
#> 
#> Efficiency factor
#> (E ) 0.6170213 
#> 
#> <<< Book >>>

book<-outdesign$book
plots<-book[,1]
dim(plots)<-c(k,s,r)
for (i in 1:r) print(t(plots[,,i]))

#>       [,1] [,2] [,3]
#>  [1,]  101  102  103
#>  [2,]  104  105  106
#>  [3,]  107  108  109
#>  [4,]  110  111  112
#>  [5,]  113  114  115
#>  [6,]  116  117  118
#>  [7,]  119  120  121
#>  [8,]  122  123  124
#>  [9,]  125  126  127
#> [10,]  128  129  130
#>       [,1] [,2] [,3]
#>  [1,]  201  202  203
#>  [2,]  204  205  206
#>  [3,]  207  208  209
#>  [4,]  210  211  212
#>  [5,]  213  214  215
#>  [6,]  216  217  218
#>  [7,]  219  220  221
#>  [8,]  222  223  224
#>  [9,]  225  226  227
#> [10,]  228  229  230

outdesign$sketch

#> $rep1
#>       [,1] [,2] [,3]
#>  [1,] "17" "5"  "24"
#>  [2,] "7"  "11" "30"
#>  [3,] "4"  "2"  "20"
#>  [4,] "26" "25" "12"
#>  [5,] "22" "18" "21"
#>  [6,] "16" "19" "15"
#>  [7,] "1"  "13" "28"
#>  [8,] "3"  "14" "6" 
#>  [9,] "8"  "10" "27"
#> [10,] "29" "9"  "23"
#> 
#> $rep2
#>       [,1] [,2] [,3]
#>  [1,] "16" "3"  "1" 
#>  [2,] "8"  "7"  "17"
#>  [3,] "21" "30" "27"
#>  [4,] "23" "15" "28"
#>  [5,] "5"  "20" "10"
#>  [6,] "11" "12" "18"
#>  [7,] "29" "2"  "24"
#>  [8,] "26" "22" "14"
#>  [9,] "19" "6"  "25"
#> [10,] "13" "9"  "4" 
#> 

# Example two 
trt<-letters[1:12]
t <- length(trt)
k<-3
r<-3
s<-t/k
outdesign<- design.alpha(trt,k,r,serie=2)

#> 
#> Alpha Design (0,1) - Serie  III 
#> 
#> Parameters Alpha Design
#> =======================
#> Treatmeans : 12
#> Block size : 3
#> Blocks     : 4
#> Replication: 3 
#> 
#> Efficiency factor
#> (E ) 0.7096774 
#> 
#> <<< Book >>>

book<-outdesign$book
plots<-book[,1]
dim(plots)<-c(k,s,r)
for (i in 1:r) print(t(plots[,,i]))

#>      [,1] [,2] [,3]
#> [1,]  101  102  103
#> [2,]  104  105  106
#> [3,]  107  108  109
#> [4,]  110  111  112
#>      [,1] [,2] [,3]
#> [1,]  201  202  203
#> [2,]  204  205  206
#> [3,]  207  208  209
#> [4,]  210  211  212
#>      [,1] [,2] [,3]
#> [1,]  301  302  303
#> [2,]  304  305  306
#> [3,]  307  308  309
#> [4,]  310  311  312

outdesign$sketch

#> $rep1
#>      [,1] [,2] [,3]
#> [1,] "g"  "c"  "k" 
#> [2,] "j"  "d"  "l" 
#> [3,] "e"  "a"  "f" 
#> [4,] "b"  "i"  "h" 
#> 
#> $rep2
#>      [,1] [,2] [,3]
#> [1,] "f"  "l"  "g" 
#> [2,] "b"  "a"  "c" 
#> [3,] "d"  "h"  "k" 
#> [4,] "e"  "j"  "i" 
#> 
#> $rep3
#>      [,1] [,2] [,3]
#> [1,] "c"  "e"  "h" 
#> [2,] "b"  "d"  "g" 
#> [3,] "i"  "l"  "a" 
#> [4,] "j"  "k"  "f" 
#>