Cyclic designs

The cyclic design is a incomplete blocks designs, it is generated from a incomplete block initial of the size k, the plan is generated and randomized. The efficient and robust cyclic designs for 6 to 30 treatments, replications <= 10.

design.cyclic(
  trt,
  k,
  r,
  serie = 2,
  rowcol = FALSE,
  seed = 0,
  kinds = "Super-Duper",
  randomization = TRUE
)

Arguments

trt
vector treatments

k
block size

r
Replications

serie
number plot, 1: 11,12; 2: 101,102; 3: 1001,1002

rowcol
TRUE: row-column design

seed
init seed random

kinds
random method

randomization
TRUE or FALSE - randomize

Value

parameters

Design parameters

sketch

Design sketch

book

Fieldbook

Details

Number o treatment 6 to 30. (r) Replication 2 to 10. (k) size of block 2 to 10. replication = i*k, "i" is value integer.

References

Kuehl, Robert(2000), Design of Experiments. 2nd ed., Duxbury. John, J.A. (1981) Efficient Cyclic Design. J. R. Statist. Soc. B, 43, No. 1, pp, 76-80.

See also

Examples

library(agricolae)
trt<-letters[1:8]
# block size = 2, replication = 6
outdesign1 <- design.cyclic(trt,k=2, r=6,serie=2)

#> 
#> cyclic design
#> Generator block basic:
#> 1 2 
#> 1 4 
#> 1 3 
#> 
#> Parameters
#> ===================
#> treatmeans : 8
#> Block size : 2
#> Replication: 6 
#> 

names(outdesign1)

#> [1] "parameters" "sketch"     "book"      

# groups 1,2,3
outdesign1$sketch[[1]]

#>      [,1] [,2]
#> [1,] "g"  "h" 
#> [2,] "a"  "h" 
#> [3,] "b"  "a" 
#> [4,] "d"  "e" 
#> [5,] "d"  "c" 
#> [6,] "e"  "f" 
#> [7,] "f"  "g" 
#> [8,] "c"  "b" 

outdesign1$sketch[[2]]

#>      [,1] [,2]
#> [1,] "g"  "d" 
#> [2,] "b"  "e" 
#> [3,] "h"  "c" 
#> [4,] "e"  "h" 
#> [5,] "c"  "f" 
#> [6,] "a"  "f" 
#> [7,] "a"  "d" 
#> [8,] "b"  "g" 

outdesign1$sketch[[3]]

#>      [,1] [,2]
#> [1,] "a"  "c" 
#> [2,] "g"  "a" 
#> [3,] "d"  "f" 
#> [4,] "h"  "f" 
#> [5,] "e"  "c" 
#> [6,] "h"  "b" 
#> [7,] "g"  "e" 
#> [8,] "d"  "b" 

outdesign1$book

#>    plots group block trt
#> 1    101     1     1   g
#> 2    102     1     1   h
#> 3    103     1     2   a
#> 4    104     1     2   h
#> 5    105     1     3   b
#> 6    106     1     3   a
#> 7    107     1     4   d
#> 8    108     1     4   e
#> 9    109     1     5   d
#> 10   110     1     5   c
#> 11   111     1     6   e
#> 12   112     1     6   f
#> 13   113     1     7   f
#> 14   114     1     7   g
#> 15   115     1     8   c
#> 16   116     1     8   b
#> 17   201     2     9   g
#> 18   202     2     9   d
#> 19   203     2    10   b
#> 20   204     2    10   e
#> 21   205     2    11   h
#> 22   206     2    11   c
#> 23   207     2    12   e
#> 24   208     2    12   h
#> 25   209     2    13   c
#> 26   210     2    13   f
#> 27   211     2    14   a
#> 28   212     2    14   f
#> 29   213     2    15   a
#> 30   214     2    15   d
#> 31   215     2    16   b
#> 32   216     2    16   g
#> 33   301     3    17   a
#> 34   302     3    17   c
#> 35   303     3    18   g
#> 36   304     3    18   a
#> 37   305     3    19   d
#> 38   306     3    19   f
#> 39   307     3    20   h
#> 40   308     3    20   f
#> 41   309     3    21   e
#> 42   310     3    21   c
#> 43   311     3    22   h
#> 44   312     3    22   b
#> 45   313     3    23   g
#> 46   314     3    23   e
#> 47   315     3    24   d
#> 48   316     3    24   b

# row-column design
outdesign2<- design.cyclic(trt,k=2, r=6, serie=2, rowcol=TRUE)

#> 
#> cyclic design
#> Generator block basic:
#> 1 2 
#> 1 4 
#> 1 3 
#> 
#> Parameters
#> ===================
#> treatmeans : 8
#> Block size : 2
#> Replication: 6 
#> 

outdesign2$sketch

#>      [,1]        
#> [1,] Character,16
#> [2,] Character,16
#> [3,] Character,16