Augmented block design

These are designs for two types of treatments: the control treatments (common) and the increased treatments. The common treatments are applied in complete randomized blocks, and the increased treatments, at random. Each treatment should be applied in any block once only. It is understood that the common treatments are of a greater interest; the standard error of the difference is much smaller than when between two increased ones in different blocks.

design.dau(
  trt1,
  trt2,
  r,
  serie = 2,
  seed = 0,
  kinds = "Super-Duper",
  name = "trt",
  randomization = TRUE
)

Arguments

trt1
checks

trt2
new

r
Replications or blocks

serie
number plot, 1: 11,12; 2: 101,102; 3: 1001,1002

seed
seed

kinds
method for to randomize

name
name of treatments

randomization
TRUE or FALSE - randomize

Value

parameters

Design parameters

book

Fieldbook

Details

kinds <- c("Wichmann-Hill", "Marsaglia-Multicarry", "Super-Duper", "Mersenne-Twister", "Knuth-TAOCP", "user-supplied", "Knuth-TAOCP-2002", "default" )

References

1. Augmented (or Hoonuiaku) Design. Federer, W.T. (1956), Hawaii Plr. rec., 55: 191-208. 2. In Augmented Designs. Federer, W.T and Raghavarao, D. (1975). Bometrics, vol. 31, No. 1 (mar.., 1975), pp. 29-35

See also

Examples

library(agricolae)
# 4 treatments and 5 blocks
T1<-c("A","B","C","D")
T2<-letters[20:26]
outdesign <-design.dau(T1,T2, r=5,serie=2)
# field book
book<-outdesign$book
by(book,book[2],function(x) paste(x[,1],"-",as.character(x[,3])))

#> block: 1
#> [1] "101 - x" "102 - B" "103 - D" "104 - A" "105 - v" "106 - C"
#> ------------------------------------------------------------ 
#> block: 2
#> [1] "201 - A" "202 - D" "203 - C" "204 - y" "205 - w" "206 - B"
#> ------------------------------------------------------------ 
#> block: 3
#> [1] "301 - u" "302 - B" "303 - A" "304 - C" "305 - D"
#> ------------------------------------------------------------ 
#> block: 4
#> [1] "401 - C" "402 - B" "403 - D" "404 - A" "405 - z"
#> ------------------------------------------------------------ 
#> block: 5
#> [1] "501 - D" "502 - t" "503 - A" "504 - B" "505 - C"

# write in hard disk
# write.table(book,"dau.txt", row.names=FALSE, sep="\t")
# file.show("dau.txt")
# Augmented designs in Completely Randomized Design
trt<-c(T1,T2)
r<-c(4,4,4,4,1,1,1,1,1,1,1)
outdesign <- design.crd(trt,r)
outdesign$book

#>    plots r trt
#> 1    101 1   D
#> 2    102 1   B
#> 3    103 1   A
#> 4    104 1   t
#> 5    105 1   C
#> 6    106 2   A
#> 7    107 1   x
#> 8    108 2   D
#> 9    109 1   v
#> 10   110 3   A
#> 11   111 2   C
#> 12   112 2   B
#> 13   113 3   C
#> 14   114 1   y
#> 15   115 4   C
#> 16   116 1   w
#> 17   117 1   u
#> 18   118 4   A
#> 19   119 3   D
#> 20   120 1   z
#> 21   121 4   D
#> 22   122 3   B
#> 23   123 4   B