Line x Tester Analysis

It makes the Line x Tester Genetic Analysis. It also estimates the general and specific combinatory ability effects and the line and tester genetic contribution.

lineXtester(replications, lines, testers, y)

Arguments

replications
Replications

lines
Lines

testers
Testers

y
Variable, response

Details

ANOVA with parents and crosses
ANOVA for line X tester analysis
ANOVA for line X tester analysis including parents
GCA Effects: Lines Effects, Testers Effects and SCA Effects.
Standard Errors for Combining Ability Effects.
Genetic Components.
...
Proportional contribution of lines, testers and their interactions to total variance

Value

return anova(formula = Y ~ Replications + Treatments).
where the Treatments contains parents, crosses and crosses vs Parents.
The crosses contains Lines, Testers and its interaction .

References

Biometrical Methods in Quantitative Genetic Analysis, Singh, Chaudhary. 1979. Hierarchial and factorial mating designs for quantitative genetic analysis in tetrasomic potato. R. Ortis; A.Golmirzaie. Theor Appl Genet (2002) 104:675-679

See also

Examples

# see structure line by testers
library(agricolae)
# example 1
data(heterosis)
site1<-subset(heterosis,heterosis[,1]==1)
output1<-with(site1,lineXtester(Replication, Female, Male, v2))

#> 
#> ANALYSIS LINE x TESTER:  v2 
#> 
#> ANOVA with parents and crosses 
#> ==============================
#>                      Df       Sum Sq     Mean Sq F value Pr(>F)
#> Replications          2  0.002674074 0.001337037   0.088 0.9159
#> Treatments           35 29.135740741 0.832449735  54.763 0.0000
#> Parents              11 21.221688889 1.929244444 126.917 0.0000
#> Parents vs. Crosses   1  1.692474074 1.692474074 111.341 0.0000
#> Crosses              23  6.221577778 0.270503382  17.795 0.0000
#> Error                70  1.064059259 0.015200847               
#> Total               107 30.202474074                           
#> 
#> ANOVA for line X tester analysis 
#> ================================
#>                 Df    Sum Sq    Mean Sq F value Pr(>F)
#> Lines            7 4.3497111 0.62138730   9.078 0.0003
#> Testers          2 0.9135444 0.45677222   6.673 0.0092
#> Lines X Testers 14 0.9583222 0.06845159   4.503 0.0000
#> Error           70 1.0640593 0.01520085               
#> 
#> ANOVA for line X tester analysis including parents 
#> ==================================================
#>                      Df       Sum Sq     Mean Sq F value Pr(>F)
#> Replications          2  0.002674074 0.001337037   0.088 0.9159
#> Treatments           35 29.135740741 0.832449735  54.763 0.0000
#> Parents              11 21.221688889 1.929244444 126.917 0.0000
#> Parents vs. Crosses   1  1.692474074 1.692474074 111.341 0.0000
#> Crosses              23  6.221577778 0.270503382  17.795 0.0000
#> Lines                 7  4.349711111 0.621387302   9.078 0.0003
#> Testers               2  0.913544444 0.456772222   6.673 0.0092
#> Lines X Testers      14  0.958322222 0.068451587   4.503 0.0000
#> Error                70  1.064059259 0.015200847               
#> Total               107 30.202474074                           
#> 
#> GCA Effects: 
#> ===========
#> Lines Effects:
#> Achirana     LT-8     MF-I    MF-II  Serrana    TPS-2   TPS-25    TPS-7 
#>    0.143   -0.346   -0.134   -0.169    0.394   -0.016    0.321   -0.194 
#> 
#> Testers Effects:
#> TPS-13 TPS-67  TS-15 
#>  0.104  0.052 -0.156 
#> 
#> SCA Effects: 
#> ===========
#>           Testers
#> Lines      TPS-13 TPS-67  TS-15
#>   Achirana -0.010  0.002  0.008
#>   LT-8      0.186 -0.042 -0.144
#>   MF-I     -0.059  0.070 -0.011
#>   MF-II    -0.021  0.055 -0.034
#>   Serrana  -0.241  0.008  0.233
#>   TPS-2    -0.044  0.101 -0.057
#>   TPS-25    0.136 -0.259  0.123
#>   TPS-7     0.054  0.064 -0.118
#> 
#> Standard Errors for Combining Ability Effects: 
#> =============================================
#> S.E. (gca for line)   : 0.04109724
#> S.E. (gca for tester) : 0.02516682
#> S.E. (sca effect)     : 0.0711825
#> S.E. (gi - gj)line    : 0.05812027
#> S.E. (gi - gj)tester  : 0.03559125
#> S.E. (sij - skl)tester: 0.1006673 
#> 
#> Genetic Components: 
#> ==================
#> Cov H.S. (line)   : 0.0614373
#> Cov H.S. (tester) : 0.01618003
#> Cov H.S. (average): 0.004651843
#> Cov F.S. (average): 0.1223343
#> F = 0, Adittive genetic variance: 0.01860737
#> F = 1, Adittive genetic variance: 0.009303686
#> F = 0, Variance due to Dominance: 0.03550049
#> F = 1, Variance due to Dominance: 0.01775025 
#> 
#> Proportional contribution of lines, testers 
#>  and their interactions to total variance 
#> ===========================================
#> Contributions of lines  : 69.91331
#> Contributions of testers: 14.68349
#> Contributions of lxt    : 15.4032 

# example 2
data(LxT)
str(LxT)

#> 'data.frame':	92 obs. of  4 variables:
#>  $ replication: int  1 2 3 4 1 2 3 4 1 2 ...
#>  $ line       : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ tester     : int  6 6 6 6 7 7 7 7 8 8 ...
#>  $ yield      : num  74.4 70.9 60.9 68 91.8 ...

output2<-with(LxT,lineXtester(replication, line, tester, yield))

#> 
#> ANALYSIS LINE x TESTER:  yield 
#> 
#> ANOVA with parents and crosses 
#> ==============================
#>                     Df      Sum Sq    Mean Sq F value Pr(>F)
#> Replications         3    83.00012   27.66671   0.304 0.8224
#> Treatments          22 32553.20239 1479.69102  16.249 0.0000
#> Parents              7  6299.88519  899.98360   9.883 0.0000
#> Parents vs. Crosses  1    53.66287   53.66287   0.589 0.4455
#> Crosses             14 26199.65433 1871.40388  20.551 0.0000
#> Error               66  6010.03298   91.06111               
#> Total               91 38646.23549                          
#> 
#> ANOVA for line X tester analysis 
#> ================================
#>                 Df    Sum Sq    Mean Sq F value Pr(>F)
#> Lines            4 10318.361 2579.59035   1.457 0.3009
#> Testers          2  1718.926  859.46289   0.485 0.6327
#> Lines X Testers  8 14162.367 1770.29589  19.441 0.0000
#> Error           66  6010.033   91.06111               
#> 
#> ANOVA for line X tester analysis including parents 
#> ==================================================
#>                     Df      Sum Sq    Mean Sq F value Pr(>F)
#> Replications         3    83.00012   27.66671   0.304 0.8224
#> Treatments          22 32553.20239 1479.69102  16.249 0.0000
#> Parents              7  6299.88519  899.98360   9.883 0.0000
#> Parents vs. Crosses  1    53.66287   53.66287   0.589 0.4455
#> Crosses             14 26199.65433 1871.40388  20.551 0.0000
#> Lines                4 10318.36140 2579.59035   1.457 0.3009
#> Testers              2  1718.92577  859.46289   0.485 0.6327
#> Lines X Testers      8 14162.36716 1770.29589  19.441 0.0000
#> Error               66  6010.03298   91.06111               
#> Total               91 38646.23549                          
#> 
#> GCA Effects: 
#> ===========
#> Lines Effects:
#>       1       2       3       4       5 
#>  -9.960  -0.718  23.817 -13.870   0.732 
#> 
#> Testers Effects:
#>      6      7      8 
#>  0.292  6.404 -6.697 
#> 
#> SCA Effects: 
#> ===========
#>      Testers
#> Lines       6       7       8
#>     1  -8.019  24.959 -16.940
#>     2 -12.546   5.717   6.828
#>     3  -9.461  -4.918  14.378
#>     4  33.136 -14.321 -18.815
#>     5  -3.111 -11.438  14.548
#> 
#> Standard Errors for Combining Ability Effects: 
#> =============================================
#> S.E. (gca for line)   : 2.75471
#> S.E. (gca for tester) : 2.133789
#> S.E. (sca effect)     : 4.771297
#> S.E. (gi - gj)line    : 3.895748
#> S.E. (gi - gj)tester  : 3.017633
#> S.E. (sij - skl)tester: 6.747633 
#> 
#> Genetic Components: 
#> ==================
#> Cov H.S. (line)   : 67.4412
#> Cov H.S. (tester) : -45.54165
#> Cov H.S. (average): 2.680894
#> Cov F.S. (average): 411.3472
#> F = 0, Adittive genetic variance: 10.72357
#> F = 1, Adittive genetic variance: 5.361787
#> F = 0, Variance due to Dominance: 839.6174
#> F = 1, Variance due to Dominance: 419.8087 
#> 
#> Proportional contribution of lines, testers 
#>  and their interactions to total variance 
#> ===========================================
#> Contributions of lines  : 39.38358
#> Contributions of testers: 6.560872
#> Contributions of lxt    : 54.05555