Splip-Plot analysis

The variance analysis of a split plot design is divided into two parts: the plot-factor analysis and the sub-plot factor analysis.

sp.plot(block, pplot, splot, Y)

Arguments

block
replications

pplot
main-plot Factor

splot
sub-plot Factor

Y
Variable, response

Value

ANOVA: Splip plot analysis

Details

The split-plot design is specifically suited for a two-factor experiment on of the factors is assigned to main plot (main-plot factor), the second factor, called the subplot factor, is assigned into subplots.

References

Statistical procedures for agricultural research. Kwanchai A. Gomez, Arturo A. Gomez. Second Edition. 1984.

See also

Examples

library(agricolae)
data(plots)
model<-with(plots,sp.plot(block,A,B,yield))

#> 
#> ANALYSIS SPLIT PLOT:  yield 
#> Class level information
#> 
#> A 	:  a1 a2 
#> B 	:  b1 b2 b3 
#> block 	:  1 2 3 
#> 
#> Number of observations:  18 
#> 
#> Analysis of Variance Table
#> 
#> Response: yield
#>       Df  Sum Sq Mean Sq F value    Pr(>F)    
#> block  2   2.111   1.056  0.5135 0.6607143    
#> A      1   0.222   0.222  0.1081 0.7735446    
#> Ea     2   4.111   2.056                      
#> B      2  29.778  14.889  3.4581 0.0827438 .  
#> A:B    2 300.444 150.222 34.8903 0.0001119 ***
#> Eb     8  34.444   4.306                      
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> cv(a) = 24.8 %, cv(b) = 35.9 %, Mean = 5.777778 
#> 

# with aov
plots[,1]<-as.factor(plots[,1])
AOV <- aov(yield ~ block + A*B + Error(block/A),data=plots)
summary(AOV)

#> 
#> Error: block
#>       Df Sum Sq Mean Sq
#> block  2  2.111   1.056
#> 
#> Error: block:A
#>           Df Sum Sq Mean Sq F value Pr(>F)
#> A          1  0.222  0.2222   0.108  0.774
#> Residuals  2  4.111  2.0556               
#> 
#> Error: Within
#>           Df Sum Sq Mean Sq F value   Pr(>F)    
#> B          2  29.78   14.89   3.458 0.082744 .  
#> A:B        2 300.44  150.22  34.890 0.000112 ***
#> Residuals  8  34.44    4.31                     
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1