Split-split-Plot analysis

The variance analysis of a split-split plot design is divided into three parts: the main-plot, subplot and sub-subplot analysis.

ssp.plot(block, pplot, splot, ssplot, Y)

Arguments

block
replications

pplot
Factor main plot

splot
Factor subplot

ssplot
Factor sub-subplot

Y
Variable, response

Details

The split-split-plot design is an extension of the split-plot design to accommodate a third factor: one factor in main-plot, other in subplot and the third factor in sub-subplot

Value

ANOVA: Splip Split plot analysis

References

Statistical procedures for agricultural research. Kwanchai A. Gomez, Arturo A. Gomez. Second Edition. 1984.

See also

Examples

# Statistical procedures for agricultural research, pag 143
# Grain Yields of Three Rice Varieties Grown under 
#Three Management practices and Five Nitrogen levels; in a
#split-split-plot design with nitrogen as main-plot, 
#management practice as subplot, and variety as sub-subplot 
#factores, with three replications.
library(agricolae)
f <- system.file("external/ssp.csv", package="agricolae")
ssp<-read.csv(f)

#> Warning: file("") only supports open = "w+" and open = "w+b": using the former

#> Error in read.table(file = file, header = header, sep = sep, quote = quote,     dec = dec, fill = fill, comment.char = comment.char, ...): no lines available in input

model<-with(ssp,ssp.plot(block,nitrogen,management,variety,yield))

#> Error in with(ssp, ssp.plot(block, nitrogen, management, variety, yield)): object 'ssp' not found

gla<-model$gl.a; glb<-model$gl.b; glc<-model$gl.c

#> Error in eval(expr, envir, enclos): object 'model' not found

#> Error in eval(expr, envir, enclos): object 'model' not found

#> Error in eval(expr, envir, enclos): object 'model' not found

Ea<-model$Ea; Eb<-model$Eb; Ec<-model$Ec

#> Error in eval(expr, envir, enclos): object 'model' not found

#> Error in eval(expr, envir, enclos): object 'model' not found

#> Error in eval(expr, envir, enclos): object 'model' not found

par(mfrow=c(1,3),cex=0.6)
out1<-with(ssp,LSD.test(yield,nitrogen,gla,Ea,console=TRUE))

#> Error in with(ssp, LSD.test(yield, nitrogen, gla, Ea, console = TRUE)): object 'ssp' not found

out2<-with(ssp,LSD.test(yield,management,glb,Eb,console=TRUE))

#> Error in with(ssp, LSD.test(yield, management, glb, Eb, console = TRUE)): object 'ssp' not found

out3<-with(ssp,LSD.test(yield,variety,glc,Ec,console=TRUE))

#> Error in with(ssp, LSD.test(yield, variety, glc, Ec, console = TRUE)): object 'ssp' not found

plot(out1,xlab="Nitrogen",las=1,variation="IQR")

#> Error in plot(out1, xlab = "Nitrogen", las = 1, variation = "IQR"): object 'out1' not found

plot(out2,xlab="Management",variation="IQR")

#> Error in plot(out2, xlab = "Management", variation = "IQR"): object 'out2' not found

plot(out3,xlab="Variety",variation="IQR")

#> Error in plot(out3, xlab = "Variety", variation = "IQR"): object 'out3' not found

# with aov
AOV<-aov(yield ~  block + nitrogen*management*variety + Error(block/nitrogen/management),data=ssp)

#> Error in terms.formula(formula, "Error", data = data): object 'ssp' not found

summary(AOV)

#> Error in summary(AOV): object 'AOV' not found