Nonparametric stability analysis

A method based on the statistical ranges of the study variable per environment for the stability analysis.

stability.nonpar(data, variable = NULL, ranking = FALSE, console = FALSE)

Arguments

data
First column the genotypes following environment

variable
Name of variable

ranking
logical, print ranking

console
logical, print output

Value

ranking

data frame

statistics

Statistical analysis chi square test

References

Haynes K G, Lambert D H, Christ B J, Weingartner D P, Douches D S, Backlund J E, Fry W and Stevenson W. 1998. Phenotypic stability of resistance to late blight in potato clones evaluated at eight sites in the United States American Journal Potato Research 75, pag 211-217.

See also

Examples

library(agricolae)
data(haynes)
stability.nonpar(haynes,"AUDPC",ranking=TRUE,console=TRUE)

#> 
#> Nonparametric Method for Stability Analysis
#> -------------------------------------------
#> 
#> Estimation and test of nonparametric measures
#> Variable: AUDPC 
#> 
#> Ranking...
#>            FL MI ME MN ND   NY PA WI
#> A84118-3    7 11 11 14  8 14.0 12 11
#> AO80432-1   6  9 13 13 12 12.0 15 14
#> AO84275-3  10 10 12  8  9  7.0 11 12
#> AWN86514-2  3  3  3  1  3  3.0  2  1
#> B0692-4     1  8  4  3  2  2.0  1  3
#> B0718-3     2  1  2  2  4  4.0  3  4
#> B0749-2F   15 12 16 10 13 13.0 13  8
#> B0767-2     4  2  1  4  1  1.0  4  2
#> Bertita     8  7  7  6 10  8.5 10  9
#> Bzura       9  4  5  5  6  5.0  5  5
#> C0083008-1 11 15 14 15 15 15.0 14 15
#> Elba       13 13 10  9 14 11.0  7 13
#> Greta       5  5  8  7  5  6.0  6  7
#> Krantz     14 16 15 16 16 16.0 16 16
#> Libertas   12  6  6 12  7  8.5  8  6
#> Stobrawa   16 14  9 11 11 10.0  9 10
#> 
#> Statistics...
#>               Mean Rank   s1   Z1    s2   Z2
#> A84118-3    741.62   13 4.82 0.22 16.70 0.34
#> AO80432-1   734.38   12 6.21 0.73 26.57 0.47
#> AO84275-3   635.88    9 6.20 0.70 28.53 0.87
#> AWN86514-2  176.88    2 5.71 0.15 23.64 0.09
#> B0692-4     224.50    4 3.11 4.37  7.12 3.28
#> B0718-3     192.50    3 6.64 1.59 30.57 1.43
#> B0749-2F    772.88   14 5.07 0.05 19.14 0.07
#> B0767-2     153.00    1 5.79 0.20 23.07 0.05
#> Bertita     502.12    8 3.57 2.73  9.43 2.30
#> Bzura       331.75    5 6.04 0.47 26.55 0.46
#> C0083008-1 1022.12   15 7.11 2.90 38.84 5.09
#> Elba        719.00   11 6.57 1.43 29.71 1.18
#> Greta       412.88    6 4.59 0.47 14.71 0.70
#> Krantz     1169.62   16 7.04 2.67 42.84 7.67
#> Libertas    500.88    7 4.50 0.59 14.00 0.87
#> Stobrawa    693.38   10 4.36 0.82 13.64 0.95
#> ------------------------
#> Sum of Z1:  20.08986
#> Sum of Z2:  25.84532
#> ------------------------
#> 
#> Test...
#> The Z-statistics are measures of stability. The test for the significance
#> of the sum of Z1 or Z2 are compared to a Chi-Square value of chi.sum. 
#> individual Z1 or Z2 are compared to a Chi-square value of chi.ind.
#> 
#>       MEAN    es1   es2      vs1      vs2  chi.ind  chi.sum
#> 1 561.4609 5.3125 21.25 1.111905 60.75223 8.733011 26.29623
#> ---
#> expectation and variance: es1, es2, vs1, vs2

# Example 2
data(CIC)
data1<-CIC$comas[,c(1,6,7,17,18)]
data2<-CIC$oxapampa[,c(1,6,7,19,20)]
cic <- rbind(data1,data2)

means <- by(cic[,5], cic[,c(2,1)], function(x) mean(x,na.rm=TRUE))
means <-as.data.frame(means[,])
cic.mean<-data.frame(genotype=row.names(means),means)
cic.mean<-delete.na(cic.mean,"greater")
out<-stability.nonpar(cic.mean)
out$ranking

#>                    Mean Rank    s1           Z1       s2           Z2
#> 762616.258   0.21156667   49 100.0 6.1406218415 5000.000 1.171984e+01
#> 762619.251   0.09073333    3  58.0 0.7113006387 1682.000 3.490158e-01
#> Atzimba      0.26909167   70 106.0 7.3524926275 5618.000 1.561060e+01
#> CC2.1        0.22568333   57  18.0 0.5089608533  162.000 4.996819e-01
#> CC2.10       0.28998333   79  68.0 1.5192700839 2312.000 1.273784e+00
#> CC2.101      0.31256667   88  12.0 0.8966861289   72.000 6.142120e-01
#> CC2.102      0.17503333   38  80.0 2.8887324335 3200.000 3.559733e+00
#> CC2.103      0.13860000   15  24.0 0.2302989456  288.000 3.591759e-01
#> CC2.104      0.28131667   76  62.0 0.9981339608 1922.000 6.330890e-01
#> CC2.105      0.43278333  104  74.0 2.1494695748 2738.000 2.226977e+00
#> CC2.106      0.19058333   45   8.0 1.2157604058   32.000 6.689042e-01
#> CC2.107      0.23630000   61  30.0 0.0607004057  450.000 2.125301e-01
#> CC2.108      0.15071667   22  46.0 0.1416363198 1058.000 3.370162e-03
#> CC2.109      0.17981667   42   2.0 1.7852579611    2.000 7.114540e-01
#> CC2.11       0.21800000   51  84.0 3.4421650993 3528.000 4.694790e+00
#> CC2.110      0.25455000   66  26.0 0.1616479470  338.000 3.098333e-01
#> CC2.111      0.27001667   71  92.0 4.6944482545 4232.000 7.660532e+00
#> CC2.112      0.26415000   68   4.0 1.5833072907    8.000 7.028391e-01
#> CC2.113      0.37413333   96  48.0 0.2062849930 1152.000 1.912771e-02
#> CC2.114      0.26607500   69  76.0 2.3837723757 2888.000 2.625580e+00
#> CC2.115      0.13686667   13  32.0 0.0284038630  512.000 1.665286e-01
#> CC2.116      0.08695000    2  38.0 0.0042231469  722.000 5.234829e-02
#> CC2.117      0.22551667   56   2.0 1.7852579611    2.000 7.114540e-01
#> CC2.118      0.28343333   77  78.0 2.6301933286 3042.000 3.068935e+00
#> CC2.119      0.12976667    9  22.0 0.3110680962  242.000 4.077896e-01
#> CC2.12       0.32376667   90  64.0 1.1597278499 2048.000 8.158386e-01
#> CC2.120      0.26336667   67  58.0 0.7113006387 1682.000 3.490158e-01
#> CC2.14       0.21693333   50  12.5 0.8602100745   78.125 6.060431e-01
#> CC2.15       0.12771667    8  28.0 0.1051151003  392.000 2.606365e-01
#> CC2.16       0.46028333  105  88.0 4.0440703729 3872.000 6.053700e+00
#> CC2.17       0.25080000   65  43.0 0.0673848450  924.500 3.126911e-03
#> CC2.18       0.39961667  103  30.0 0.0607004057  450.000 2.125301e-01
#> CC2.19       0.15001667   21  70.0 1.7172184289 2450.000 1.553598e+00
#> CC2.2        0.36513333   94  18.0 0.5089608533  162.000 4.996819e-01
#> CC2.20       0.37440000   97  84.0 3.4421650993 3528.000 4.694790e+00
#> CC2.21       0.29431667   80   0.0 1.9993267835    0.000 7.143373e-01
#> CC2.22       0.30085000   83  66.0 1.3334398909 2178.000 1.028646e+00
#> CC2.23       0.30753333   86  52.0 0.3719367953 1352.000 9.551027e-02
#> CC2.24       0.16428333   31  76.0 2.3837723757 2888.000 2.625580e+00
#> CC2.26       0.27376667   73  34.0 0.0082254723  578.000 1.237164e-01
#> CC2.27       0.16003333   27  32.0 0.0284038630  512.000 1.665286e-01
#> CC2.28       0.16395000   30  22.0 0.3110680962  242.000 4.077896e-01
#> CC2.3        0.30043333   82 106.0 7.3524926275 5618.000 1.561060e+01
#> CC2.30       0.30310000   84 102.0 6.5324606182 5202.000 1.293033e+01
#> CC2.31       0.36036667   92  94.0 5.0378144233 4418.000 8.564747e+00
#> CC2.32       0.39414167  101  38.0 0.0042231469  722.000 5.234829e-02
#> CC2.33       0.13773333   14  26.0 0.1616479470  338.000 3.098333e-01
#> CC2.34       0.13273333   11   6.0 1.3934747723   18.000 6.885975e-01
#> CC2.35       0.36385000   93  86.0 3.7370586601 3698.000 5.344785e+00
#> CC2.36       0.23273333   59  46.0 0.1416363198 1058.000 3.370162e-03
#> CC2.37       0.27213333   72 102.0 6.5324606182 5202.000 1.293033e+01
#> CC2.38       0.21893333   52  94.0 5.0378144233 4418.000 8.564747e+00
#> CC2.39       0.39881667  102 100.0 6.1406218415 5000.000 1.171984e+01
#> CC2.4        0.29800000   81 104.0 6.9364175469 5408.000 1.422605e+01
#> CC2.40       0.20903333   48  14.0 0.7553262184   98.000 5.799127e-01
#> CC2.41       0.15696667   26  74.0 2.1494695748 2738.000 2.226977e+00
#> CC2.43       0.49376667  106  50.0 0.2830518182 1250.000 4.926966e-02
#> CC2.44       0.11718333    4  70.0 1.7172184289 2450.000 1.553598e+00
#> CC2.46       0.32461667   91   8.0 1.2157604058   32.000 6.689042e-01
#> CC2.47       0.14420000   16  16.0 0.6260844598  128.000 5.415611e-01
#> CC2.48       0.23183333   58  42.0 0.0486934294  882.000 8.501204e-03
#> CC2.49       0.27411667   74  56.0 0.5860612056 1568.000 2.434942e-01
#> CC2.5        0.31270000   89  62.0 0.9981339608 1922.000 6.330890e-01
#> CC2.50       0.24668333   64  50.0 0.2830518182 1250.000 4.926966e-02
#> CC2.51       0.13461667   12   4.0 1.5833072907    8.000 7.028391e-01
#> CC2.52       0.11930000    7  13.5 0.7895301192   91.125 5.888864e-01
#> CC2.53       0.24228333   63  54.0 0.4729399245 1458.000 1.596337e-01
#> CC2.54       0.38315000  100  96.0 5.3932987441 4608.000 9.540476e+00
#> CC2.55       0.17501667   37  40.0 0.0203992122  800.000 2.631117e-02
#> CC2.56       0.18508333   43  10.0 1.0501641913   50.000 6.440041e-01
#> CC2.58       0.23298333   60  82.0 3.1593896904 3362.000 4.100740e+00
#> CC2.6        0.16106667   28  96.0 5.3932987441 4608.000 9.540476e+00
#> CC2.60       0.17676667   40  92.0 4.6944482545 4232.000 7.660532e+00
#> CC2.61       0.14451667   17  36.0 0.0001652336  648.000 8.524818e-02
#> CC2.62       0.16963333   33  10.0 1.0501641913   50.000 6.440041e-01
#> CC2.63       0.38006667   99  86.0 3.7370586601 3698.000 5.344785e+00
#> CC2.64       0.18548333   44  20.0 0.4039553987  200.000 4.548698e-01
#> CC2.65       0.16696667   32  82.0 3.1593896904 3362.000 4.100740e+00
#> CC2.66       0.31165000   87 104.0 6.9364175469 5408.000 1.422605e+01
#> CC2.68       0.14458333   18  44.0 0.0891057986  968.000 3.527591e-04
#> CC2.69       0.11881667    5  28.0 0.1051151003  392.000 2.606365e-01
#> CC2.7        0.52773333  108  60.0 0.8486582238 1800.000 4.781927e-01
#> CC2.70       0.17510000   39  90.0 4.3632002378 4050.000 6.824577e+00
#> CC2.71       0.22300000   54 108.0 7.7806858602 5832.000 1.708766e+01
#> CC2.72       0.17151667   34  68.0 1.5192700839 2312.000 1.273784e+00
#> CC2.73       0.17363333   36  66.0 1.3334398909 2178.000 1.028646e+00
#> CC2.74       0.20510000   47  24.0 0.2302989456  288.000 3.591759e-01
#> CC2.75       0.30496667   85  34.0 0.0082254723  578.000 1.237164e-01
#> CC2.76       0.19255000   46  98.0 5.7609012168 4802.000 1.059104e+01
#> CC2.77       0.17205000   35  72.0 1.9272849258 2592.000 1.870501e+00
#> CC2.8        0.17890000   41  72.0 1.9272849258 2592.000 1.870501e+00
#> CC2.83       0.15110000   23  56.0 0.5860612056 1568.000 2.434942e-01
#> CC2.84       0.22048333   53  48.0 0.2062849930 1152.000 1.912771e-02
#> CC2.85       0.15310000   24  52.0 0.3719367953 1352.000 9.551027e-02
#> CC2.86       0.13200000   10   6.0 1.3934747723   18.000 6.885975e-01
#> CC2.88       0.14791667   20  54.0 0.4729399245 1458.000 1.596337e-01
#> CC2.9        0.27691667   75  40.0 0.0203992122  800.000 2.631117e-02
#> CC2.90       0.37865000   98  60.0 0.8486582238 1800.000 4.781927e-01
#> CC2.91       0.14760000   19  64.0 1.1597278499 2048.000 8.158386e-01
#> CC2.92       0.11891667    6  20.0 0.4039553987  200.000 4.548698e-01
#> CC2.94       0.16336667   29  36.0 0.0001652336  648.000 8.524818e-02
#> CC2.96       0.23653333   62  43.0 0.0673848450  924.500 3.126911e-03
#> CC2.97       0.15518333   25  78.0 2.6301933286 3042.000 3.068935e+00
#> CC2.98       0.36918333   95  88.0 4.0440703729 3872.000 6.053700e+00
#> CC2.99       0.22388333   55  16.0 0.6260844598  128.000 5.415611e-01
#> Chata_Blanca 0.70826667  109  80.0 2.8887324335 3200.000 3.559733e+00
#> Lbr_40       0.08095000    1 108.0 7.7806858602 5832.000 1.708766e+01
#> Monserrate   0.28824167   78  90.0 4.3632002378 4050.000 6.824577e+00
#> Yungay       0.51476667  107  98.0 5.7609012168 4802.000 1.059104e+01

out$statistics

#>        MEAN      es1 es2      vs1     vs2  chi.ind  chi.sum
#> 1 0.2425566 36.33028 990 660.1667 1372041 12.27643 134.3688