Stability analysis. SHUKLA'S STABILITY VARIANCE AND KANG'S

This procedure calculates the stability variations as well as the statistics of selection for the yield and the stability. The averages of the genotype through the different environment repetitions are required for the calculations. The mean square error must be calculated from the joint variance analysis.

stability.par(
  data,
  rep,
  MSerror,
  alpha = 0.1,
  main = NULL,
  cova = FALSE,
  name.cov = NULL,
  file.cov = 0,
  console = FALSE
)

Arguments

data
matrix of averages, by rows the genotypes and columns the environment

rep
Number of repetitions

MSerror
Mean Square Error

alpha
Label significant

main
Title

cova
Covariable

name.cov
Name covariable

file.cov
Data covariable

console
logical, print output

Value

analysis

Analysis of variance

statistics

Statistics of the model

stability

summary stability analysis

Details

Stable (i) determines the contribution of each genotype to GE interaction by calculating var(i); (ii) assigns ranks to genotypes from highest to lowest yield receiving the rank of 1; (iii) calculates protected LSD for mean yield comparisons; (iv) adjusts yield rank according to LSD (the adjusted rank labeled Y); (v) determines significance of var(i) usign an aproximate F-test; (vi) assigns stability rating (S) as follows: -8, -4 and -2 for var(i) significant at the 0.01, 0.05 and 0.10 probability levels, and 0 for nonsignificant var(i) ( the higher the var(i), the less stable the genotype); (vii) sums adjusted yield rank, Y, and stability rating, S, for each genotype to determine YS(i) statistic; and (viii) calculates mean YS(i) and identifies genotypes (selection) with YS(i) > mean YS(i).

References

Kang, M. S. 1993. Simultaneous selection for yield and stability: Consequences for growers. Agron. J. 85:754-757. Manjit S. Kang and Robert Mangari. 1995. Stable: A basic program for calculating stability and yield-stability statistics. Agron. J. 87:276-277

See also

Examples

library(agricolae)
# example 1
# Experimental data,
# replication rep= 4
# Mean square error, MSerror = 1.8
# 12 environment
# 17 genotype  = 1,2,3,.., 17
# yield averages of 13 genotypes in localities
f <- system.file("external/dataStb.csv", package="agricolae")
dataStb<-read.csv(f)
stability.par(dataStb, rep=4, MSerror=1.8, alpha=0.1, main="Genotype",console=TRUE)

#> 
#>  INTERACTIVE PROGRAM FOR CALCULATING SHUKLA'S STABILITY VARIANCE AND KANG'S
#>                         YIELD - STABILITY (YSi) STATISTICS
#>  Genotype 
#>  Environmental index  - covariate 
#> 
#>  Analysis of Variance
#> 
#>                Df    Sum Sq  Mean Sq F value Pr(>F)
#> Total         203 2964.1716                        
#> Genotypes      16  186.9082  11.6818    4.17 <0.001
#> Environments   11 2284.0116 207.6374  115.35 <0.001
#> Interaction   176  493.2518   2.8026    1.56 <0.001
#> Heterogeneity  16   44.8576   2.8036       1  0.459
#> Residual      160  448.3942   2.8025    1.56 <0.001
#> Pooled Error  576                1.8               
#> 
#>  Genotype. Stability statistics
#> 
#>       Mean Sigma-square  . s-square  . Ecovalence
#> A 7.383333     2.474081 ns 2.449076 ns  25.826563
#> B 6.783333     1.600869 ns 1.434734 ns  17.351269
#> C 7.250000     0.567657 ns 0.633936 ns   7.323033
#> D 6.783333     2.611778 ns 2.134731 ns  27.163033
#> E 7.066667     1.862364 ns 2.047627 ns  19.889308
#> F 6.916667     3.575818  * 3.951442  *  36.519896
#> G 7.808333     3.580929  * 3.957319  *  36.569504
#> H 7.908333     2.723717 ns 2.118116 ns  28.249504
#> I 7.275000     4.248566 ** 3.936130  *  43.049504
#> J 7.083333     2.273838 ns 2.506382 ns  23.883033
#> K 6.433333     2.560384 ns 2.551518 ns  26.664210
#> L 6.891667     1.558061 ns 1.732557 ns  16.935779
#> M 6.791667     3.483879  * 3.275985 ns  35.627543
#> N 7.491667     5.164848 ** 4.875189 **  51.942837
#> O 7.658333     2.380202 ns 2.635025 ns  24.915386
#> P 6.425000     3.445414  * 3.713885  *  35.254210
#> Q 6.158333     3.531232  * 3.688232  *  36.087151
#> 
#> 
#> Signif. codes:  0 '**' 0.01 '*' 0.05 'ns' 1
#> 
#> Simultaneous selection for yield and stability  (++)
#> 
#>      Yield Rank Adj.rank Adjusted Stab.var Stab.rating YSi ...
#> A 7.383333   13        1       14 2.474081           0  14   +
#> B 6.783333    4       -1        3 1.600869           0   3    
#> C 7.250000   11        1       12 0.567657           0  12   +
#> D 6.783333    4       -1        3 2.611778           0   3    
#> E 7.066667    9        1       10 1.862364           0  10   +
#> F 6.916667    8       -1        7 3.575818          -4   3    
#> G 7.808333   16        2       18 3.580929          -4  14   +
#> H 7.908333   17        2       19 2.723717           0  19   +
#> I 7.275000   12        1       13 4.248566          -8   5    
#> J 7.083333   10        1       11 2.273838           0  11   +
#> K 6.433333    3       -2        1 2.560384           0   1    
#> L 6.891667    7       -1        6 1.558061           0   6    
#> M 6.791667    6       -1        5 3.483879          -4   1    
#> N 7.491667   14        1       15 5.164848          -8   7   +
#> O 7.658333   15        2       17 2.380202           0  17   +
#> P 6.425000    2       -2        0 3.445414          -4  -4    
#> Q 6.158333    1       -3       -2 3.531232          -4  -6    
#> 
#>  Yield Mean: 7.065196
#>  YS    Mean: 6.823529
#>  LSD (0.05): 0.4511874
#>  - - - - - - - - - - -
#>  +   selected genotype
#>  ++  Reference: Kang, M. S. 1993. Simultaneous selection for yield
#>  and stability: Consequences for growers. Agron. J. 85:754-757.

#example 2 covariable. precipitation
precipitation<- c(1000,1100,1200,1300,1400,1500,1600,1700,1800,1900,2000,2100)
stability.par(dataStb, rep=4, MSerror=1.8, alpha=0.1, main="Genotype",
 cova=TRUE, name.cov="Precipitation", file.cov=precipitation,console=TRUE)

#> 
#>  INTERACTIVE PROGRAM FOR CALCULATING SHUKLA'S STABILITY VARIANCE AND KANG'S
#>                         YIELD - STABILITY (YSi) STATISTICS
#>  Genotype 
#>  Precipitation  - covariate 
#> 
#>  Analysis of Variance
#> 
#>                Df    Sum Sq  Mean Sq F value Pr(>F)
#> Total         203 2964.1716                        
#> Genotypes      16  186.9082  11.6818    4.17 <0.001
#> Environments   11 2284.0116 207.6374  115.35 <0.001
#> Interaction   176  493.2518   2.8026    1.56 <0.001
#> Heterogeneity  16    4.3577   0.2724    0.09      1
#> Residual      160   488.894   3.0556     1.7 <0.001
#> Pooled Error  576                1.8               
#> 
#>  Genotype. Stability statistics
#> 
#>       Mean Sigma-square  . s-square  . Ecovalence
#> A 7.383333     2.474081 ns 2.714197 ns  25.826563
#> B 6.783333     1.600869 ns 1.758725 ns  17.351269
#> C 7.250000     0.567657 ns 0.624604 ns   7.323033
#> D 6.783333     2.611778 ns 2.852117 ns  27.163033
#> E 7.066667     1.862364 ns 2.030613 ns  19.889308
#> F 6.916667     3.575818  * 3.934039  *  36.519896
#> G 7.808333     3.580929  * 3.933173  *  36.569504
#> H 7.908333     2.723717 ns 2.960728 ns  28.249504
#> I 7.275000     4.248566 ** 4.656603 **  43.049504
#> J 7.083333     2.273838 ns 2.473687 ns  23.883033
#> K 6.433333     2.560384 ns 2.791349 ns  26.664210
#> L 6.891667     1.558061 ns 1.713683 ns  16.935779
#> M 6.791667     3.483879  * 3.741592  *  35.627543
#> N 7.491667     5.164848 ** 5.683086 **  51.942837
#> O 7.658333     2.380202 ns 2.611730 ns  24.915386
#> P 6.425000     3.445414  * 3.687730  *  35.254210
#> Q 6.158333     3.531232  * 3.777338  *  36.087151
#> 
#> 
#> Signif. codes:  0 '**' 0.01 '*' 0.05 'ns' 1
#> 
#> Simultaneous selection for yield and stability  (++)
#> 
#>      Yield Rank Adj.rank Adjusted Stab.var Stab.rating YSi ...
#> A 7.383333   13        1       14 2.474081           0  14   +
#> B 6.783333    4       -1        3 1.600869           0   3    
#> C 7.250000   11        1       12 0.567657           0  12   +
#> D 6.783333    4       -1        3 2.611778           0   3    
#> E 7.066667    9        1       10 1.862364           0  10   +
#> F 6.916667    8       -1        7 3.575818          -4   3    
#> G 7.808333   16        2       18 3.580929          -4  14   +
#> H 7.908333   17        2       19 2.723717           0  19   +
#> I 7.275000   12        1       13 4.248566          -8   5    
#> J 7.083333   10        1       11 2.273838           0  11   +
#> K 6.433333    3       -2        1 2.560384           0   1    
#> L 6.891667    7       -1        6 1.558061           0   6    
#> M 6.791667    6       -1        5 3.483879          -4   1    
#> N 7.491667   14        1       15 5.164848          -8   7   +
#> O 7.658333   15        2       17 2.380202           0  17   +
#> P 6.425000    2       -2        0 3.445414          -4  -4    
#> Q 6.158333    1       -3       -2 3.531232          -4  -6    
#> 
#>  Yield Mean: 7.065196
#>  YS    Mean: 6.823529
#>  LSD (0.05): 0.4511874
#>  - - - - - - - - - - -
#>  +   selected genotype
#>  ++  Reference: Kang, M. S. 1993. Simultaneous selection for yield
#>  and stability: Consequences for growers. Agron. J. 85:754-757.