frequency Table of a Histogram

It finds the absolute, relative and accumulated frequencies with the class intervals defined from a previously calculated histogram by the "hist" of R function.

table.freq(object)

Arguments

object
Object by function graph.freq()

Value

Frequency table.

Lower

Lower limit class

Upper

Upper limit class

Main

class point

Frequency

Frequency

Percentage

Percentage frequency

CF

Cumulative frequency

CPF

Cumulative Percentage frequency

See also

Examples

library(agricolae)
data(growth)
h2<-with(growth,graph.freq(height,plot=FALSE))
print(table.freq(h2),row.names=FALSE)

#>  Lower Upper Main Frequency Percentage CF   CPF
#>    6.0   7.2  6.6         1        3.3  1   3.3
#>    7.2   8.4  7.8         0        0.0  1   3.3
#>    8.4   9.6  9.0         4       13.3  5  16.7
#>    9.6  10.8 10.2         8       26.7 13  43.3
#>   10.8  12.0 11.4        11       36.7 24  80.0
#>   12.0  13.2 12.6         6       20.0 30 100.0